Base Rate

This function allows you to obtain current Base Rates and Base Lending Rates for retail loans or financing facilities and Indicative Effective Lending Rates for a standard housing loan/home financing offered by financial institutions from the BNM API.

base_rate(bank_code = NULL)

Arguments

bank_code
8 characters of SWIFT code per https://api.bnm.gov.my/

Source

https://api.bnm.gov.my/

Examples

base_rate()

#>    bank_code                                                 bank_name
#> 1   BKKBMYKL                                       Bangkok Bank Berhad
#> 2   CIBBMYKL                                          CIMB Bank Berhad
#> 3   CITIMYKL                                           Citibank Berhad
#> 4   HLBBMYKL                           Hong Leong Bank Malaysia Berhad
#> 5   HBMBMYKL                                 HSBC Bank Malaysia Berhad
#> 6   ICBKMYKL Industrial and Commercial Bank of China (Malaysia) Berhad
#> 7   MBBEMYKL                                    Malayan Banking Berhad
#> 8   OCBCMYKL                               OCBC Bank (Malaysia) Berhad
#> 9   PBBEMYKL                                        Public Bank Berhad
#> 10  RHBBMYKL                                           RHB Bank Berhad
#> 11  SCBLMYKX                   Standard Chartered Bank Malaysia Berhad
#> 12  UOVBMYKL                      United Overseas Bank (Malaysia) Bhd.
#> 13  AIBBMYKL                                 Affin Islamic Bank Berhad
#> 14  ALSRMYKL                              Alliance Islamic Bank Berhad
#> 15  AISLMYKL                                     AmBank Islamic Berhad
#> 16  BIMBMYKL                                Bank Islam Malaysia Berhad
#> 17  BMMBMYKL                             Bank Muamalat Malaysia Berhad
#> 18  CTBBMYKL                                  CIMB Islamic Bank Berhad
#> 19  HLIBMYKL                            Hong Leong Islamic Bank Berhad
#> 20  HMABMYKL                               HSBC Amanah Malaysia Berhad
#> 21  KFHOMYKL                    Kuwait Finance House (Malaysia) Berhad
#> 22  MBISMYKL                                    Maybank Islamic Berhad
#> 23  AFBQMYKL                                          MBSB Bank Berhad
#> 24  OABBMYKL                                  OCBC Al-Amin Bank Berhad
#> 25  PUIBMYKL                                Public Islamic Bank Berhad
#> 26  RHBAMYKL                                   RHB Islamic Bank Berhad
#> 27  SCSRMYKK                          Standard Chartered Saadiq Berhad
#> 28  AGOBMYKL                                                  Agrobank
#> 29  BSNAMYK1                                    Bank Simpanan Nasional
#> 30  PHBMMYKL                                         Affin Bank Berhad
#> 31  MFBBMYKL                             Alliance Bank Malaysia Berhad
#> 32  ARBKMYKL                                         AmBank (M) Berhad
#> 33  BKCHMYKL                                  Bank of China (M) Berhad
#> 34  RJHIMYKL                  Al Rajhi Banking & Investment (M) Berhad
#> 35  BKRMMYKL                          Bank Kerjasama Rakyat (M) Berhad
#>    base_rate base_lending_rate indicative_eff_lending_rate base_financing_rate
#> 1       4.47              7.14                        5.67                  NA
#> 2       4.00              6.85                        4.75                  NA
#> 3       3.65              6.80                        4.45                  NA
#> 4       3.88              6.89                        4.75                  NA
#> 5       3.64              6.74                        4.75                  NA
#> 6       3.77              6.70                        4.60                  NA
#> 7       3.00              6.65                        4.35                  NA
#> 8       3.83              6.76                        4.70                  NA
#> 9       3.52              6.72                        4.35                  NA
#> 10      3.75              6.70                        4.75                  NA
#> 11      3.52              6.70                        4.52                  NA
#> 12      3.86              6.82                        4.61                  NA
#> 13      3.95                NA                        4.76                6.81
#> 14      3.82                NA                        4.36                6.67
#> 15      3.85                NA                        4.50                6.70
#> 16      3.77                NA                        4.57                6.72
#> 17      3.81                NA                        4.81                6.81
#> 18      4.00                NA                        4.75                6.85
#> 19      3.88                NA                        4.60                6.89
#> 20      3.64                NA                        4.75                6.74
#> 21      3.50                NA                        4.45                7.39
#> 22      3.00                NA                        4.35                6.65
#> 23      3.90                NA                        4.45                6.75
#> 24      3.83                NA                        4.70                6.76
#> 25      3.52                NA                        4.35                6.72
#> 26      3.75                NA                        4.75                6.70
#> 27      3.52                NA                        4.52                6.70
#> 28      3.60                NA                          NA                6.75
#> 29      3.85                NA                        4.45                6.60
#> 30      3.95              6.81                        4.76                  NA
#> 31      3.82              6.67                        4.36                  NA
#> 32      3.85              6.70                        4.50                  NA
#> 33      3.80              6.60                        4.80                  NA
#> 34      4.10                NA                        5.45                7.00
#> 35      3.85                NA                        4.65                6.83