Consumer Alert

This function allows you to obtain a list of known companies and websites which are neither authorised nor approved under the relevant laws and regulations administered by BNM from the BNM API.

Usage

get_consumer_alert()

Source

https://api.bnm.gov.my/

Examples

get_consumer_alert()
#> # A tibble: 431 x 4
#>    name                                   regisration_number added_date websites
#>    <chr>                                  <chr>              <chr>      <list>  
#>  1 1globalcash                            ""                 2012-07-13 <chr [0~
#>  2 1Gold.com.my                           ""                 2012-07-13 <chr [1~
#>  3 3Sixty Venture Capital PLC             ""                 2014-12-30 <chr [1~
#>  4 A.A.M Global Corporation Sdn Bhd       ""                 2017-05-17 <chr [0~
#>  5 Ace Global Sales & Services            ""                 2013-05-02 <chr [0~
#>  6 Ace Dimension Network Sdn Bhd          ""                 2015-04-10 <chr [0~
#>  7 AE Group Holding Pte. Ltd. (201322498~ ""                 2015-05-14 <chr [1~
#>  8 Agarwood Venture   (002273031-A)       ""                 2014-02-19 <chr [0~
#>  9 Agar Wood Chamber of Commerce Malaysia ""                 2015-05-21 <chr [0~
#> 10 Ahmad Zulkhairi Associates PLT (LLP00~ ""                 2017-06-22 <chr [1~
#> # ... with 421 more rows