Exchange Rates: KL Interbank Foreign Exchange Market

Obtain currency exchange rates from the Interbank Foreign Exchange Market in Kuala Lumpur in relation to Ringgit.

Usage

get_exchange_rate(currency = NULL, session = "1130", quote = "rm")

Source

https://apikijangportal.bnm.gov.my/

Arguments

currency

3-characters currency code based on ISO4217 standard

session

Character string "0900", "1130", "1200" or "1700". A snapshot of the exchange rate daily at 0900, 1130, 1200 and 1700 intervals

quote

Base currency (Ringgit ("rm") or foreign currency ("fx")) as the denominator for the exchange rate

Examples

get_exchange_rate()
#> [[1]]
#> [[1]]$currency_code
#> [1] "CHF"
#> 
#> [[1]]$unit
#> [1] 1
#> 
#> [[1]]$rate
#> [[1]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[1]]$rate$buying_rate
#> [1] 5.2697
#> 
#> [[1]]$rate$selling_rate
#> [1] 5.2755
#> 
#> [[1]]$rate$middle_rate
#> [1] 5.2726
#> 
#> 
#> 
#> [[2]]
#> [[2]]$currency_code
#> [1] "CAD"
#> 
#> [[2]]$unit
#> [1] 1
#> 
#> [[2]]$rate
#> [[2]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[2]]$rate$buying_rate
#> [1] 3.4037
#> 
#> [[2]]$rate$selling_rate
#> [1] 3.4068
#> 
#> [[2]]$rate$middle_rate
#> [1] 3.4052
#> 
#> 
#> 
#> [[3]]
#> [[3]]$currency_code
#> [1] "BND"
#> 
#> [[3]]$unit
#> [1] 1
#> 
#> [[3]]$rate
#> [[3]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[3]]$rate$buying_rate
#> [1] 3.4747
#> 
#> [[3]]$rate$selling_rate
#> [1] 3.4782
#> 
#> [[3]]$rate$middle_rate
#> [1] 3.4764
#> 
#> 
#> 
#> [[4]]
#> [[4]]$currency_code
#> [1] "AUD"
#> 
#> [[4]]$unit
#> [1] 1
#> 
#> [[4]]$rate
#> [[4]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[4]]$rate$buying_rate
#> [1] 3.0301
#> 
#> [[4]]$rate$selling_rate
#> [1] 3.0344
#> 
#> [[4]]$rate$middle_rate
#> [1] 3.0323
#> 
#> 
#> 
#> [[5]]
#> [[5]]$currency_code
#> [1] "AED"
#> 
#> [[5]]$unit
#> [1] 100
#> 
#> [[5]]$rate
#> [[5]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[5]]$rate$buying_rate
#> [1] 127.3891
#> 
#> [[5]]$rate$selling_rate
#> [1] 127.4777
#> 
#> [[5]]$rate$middle_rate
#> [1] 127.4334
#> 
#> 
#> 
#> [[6]]
#> [[6]]$currency_code
#> [1] "CNY"
#> 
#> [[6]]$unit
#> [1] 1
#> 
#> [[6]]$rate
#> [[6]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[6]]$rate$buying_rate
#> [1] 0.6462
#> 
#> [[6]]$rate$selling_rate
#> [1] 0.6467
#> 
#> [[6]]$rate$middle_rate
#> [1] 0.6465
#> 
#> 
#> 
#> [[7]]
#> [[7]]$currency_code
#> [1] "NPR"
#> 
#> [[7]]$unit
#> [1] 100
#> 
#> [[7]]$rate
#> [[7]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[7]]$rate$buying_rate
#> [1] 3.512
#> 
#> [[7]]$rate$selling_rate
#> [1] 3.5145
#> 
#> [[7]]$rate$middle_rate
#> [1] 3.5132
#> 
#> 
#> 
#> [[8]]
#> [[8]]$currency_code
#> [1] "SDR"
#> 
#> [[8]]$unit
#> [1] 1
#> 
#> [[8]]$rate
#> [[8]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[8]]$rate$buying_rate
#> NULL
#> 
#> [[8]]$rate$selling_rate
#> NULL
#> 
#> [[8]]$rate$middle_rate
#> [1] 6.1997
#> 
#> 
#> 
#> [[9]]
#> [[9]]$currency_code
#> [1] "VND"
#> 
#> [[9]]$unit
#> [1] 100
#> 
#> [[9]]$rate
#> [[9]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[9]]$rate$buying_rate
#> [1] 0.0193
#> 
#> [[9]]$rate$selling_rate
#> [1] 0.0193
#> 
#> [[9]]$rate$middle_rate
#> [1] 0.0193
#> 
#> 
#> 
#> [[10]]
#> [[10]]$currency_code
#> [1] "USD"
#> 
#> [[10]]$unit
#> [1] 1
#> 
#> [[10]]$rate
#> [[10]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[10]]$rate$buying_rate
#> [1] 4.679
#> 
#> [[10]]$rate$selling_rate
#> [1] 4.682
#> 
#> [[10]]$rate$middle_rate
#> [1] 4.6805
#> 
#> 
#> 
#> [[11]]
#> [[11]]$currency_code
#> [1] "TWD"
#> 
#> [[11]]$unit
#> [1] 100
#> 
#> [[11]]$rate
#> [[11]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[11]]$rate$buying_rate
#> [1] 14.7064
#> 
#> [[11]]$rate$selling_rate
#> [1] 14.7205
#> 
#> [[11]]$rate$middle_rate
#> [1] 14.7135
#> 
#> 
#> 
#> [[12]]
#> [[12]]$currency_code
#> [1] "THB"
#> 
#> [[12]]$unit
#> [1] 100
#> 
#> [[12]]$rate
#> [[12]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[12]]$rate$buying_rate
#> [1] 13.3172
#> 
#> [[12]]$rate$selling_rate
#> [1] 13.3352
#> 
#> [[12]]$rate$middle_rate
#> [1] 13.3262
#> 
#> 
#> 
#> [[13]]
#> [[13]]$currency_code
#> [1] "SGD"
#> 
#> [[13]]$unit
#> [1] 1
#> 
#> [[13]]$rate
#> [[13]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[13]]$rate$buying_rate
#> [1] 3.4744
#> 
#> [[13]]$rate$selling_rate
#> [1] 3.4785
#> 
#> [[13]]$rate$middle_rate
#> [1] 3.4764
#> 
#> 
#> 
#> [[14]]
#> [[14]]$currency_code
#> [1] "SAR"
#> 
#> [[14]]$unit
#> [1] 100
#> 
#> [[14]]$rate
#> [[14]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[14]]$rate$buying_rate
#> [1] 124.76
#> 
#> [[14]]$rate$selling_rate
#> [1] 124.8467
#> 
#> [[14]]$rate$middle_rate
#> [1] 124.8034
#> 
#> 
#> 
#> [[15]]
#> [[15]]$currency_code
#> [1] "PKR"
#> 
#> [[15]]$unit
#> [1] 100
#> 
#> [[15]]$rate
#> [[15]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[15]]$rate$buying_rate
#> [1] 1.6317
#> 
#> [[15]]$rate$selling_rate
#> [1] 1.6356
#> 
#> [[15]]$rate$middle_rate
#> [1] 1.6337
#> 
#> 
#> 
#> [[16]]
#> [[16]]$currency_code
#> [1] "PHP"
#> 
#> [[16]]$unit
#> [1] 100
#> 
#> [[16]]$rate
#> [[16]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[16]]$rate$buying_rate
#> [1] 8.4064
#> 
#> [[16]]$rate$selling_rate
#> [1] 8.4148
#> 
#> [[16]]$rate$middle_rate
#> [1] 8.4106
#> 
#> 
#> 
#> [[17]]
#> [[17]]$currency_code
#> [1] "NZD"
#> 
#> [[17]]$unit
#> [1] 1
#> 
#> [[17]]$rate
#> [[17]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[17]]$rate$buying_rate
#> [1] 2.791
#> 
#> [[17]]$rate$selling_rate
#> [1] 2.7947
#> 
#> [[17]]$rate$middle_rate
#> [1] 2.7929
#> 
#> 
#> 
#> [[18]]
#> [[18]]$currency_code
#> [1] "MMK"
#> 
#> [[18]]$unit
#> [1] 100
#> 
#> [[18]]$rate
#> [[18]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[18]]$rate$buying_rate
#> [1] 0.2235
#> 
#> [[18]]$rate$selling_rate
#> [1] 0.2236
#> 
#> [[18]]$rate$middle_rate
#> [1] 0.2235
#> 
#> 
#> 
#> [[19]]
#> [[19]]$currency_code
#> [1] "KRW"
#> 
#> [[19]]$unit
#> [1] 100
#> 
#> [[19]]$rate
#> [[19]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[19]]$rate$buying_rate
#> [1] 0.3612
#> 
#> [[19]]$rate$selling_rate
#> [1] 0.3616
#> 
#> [[19]]$rate$middle_rate
#> [1] 0.3614
#> 
#> 
#> 
#> [[20]]
#> [[20]]$currency_code
#> [1] "KHR"
#> 
#> [[20]]$unit
#> [1] 100
#> 
#> [[20]]$rate
#> [[20]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[20]]$rate$buying_rate
#> [1] 0.1126
#> 
#> [[20]]$rate$selling_rate
#> [1] 0.1139
#> 
#> [[20]]$rate$middle_rate
#> [1] 0.1133
#> 
#> 
#> 
#> [[21]]
#> [[21]]$currency_code
#> [1] "JPY"
#> 
#> [[21]]$unit
#> [1] 100
#> 
#> [[21]]$rate
#> [[21]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[21]]$rate$buying_rate
#> [1] 3.1235
#> 
#> [[21]]$rate$selling_rate
#> [1] 3.1261
#> 
#> [[21]]$rate$middle_rate
#> [1] 3.1248
#> 
#> 
#> 
#> [[22]]
#> [[22]]$currency_code
#> [1] "INR"
#> 
#> [[22]]$unit
#> [1] 100
#> 
#> [[22]]$rate
#> [[22]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[22]]$rate$buying_rate
#> [1] 5.6166
#> 
#> [[22]]$rate$selling_rate
#> [1] 5.624
#> 
#> [[22]]$rate$middle_rate
#> [1] 5.6203
#> 
#> 
#> 
#> [[23]]
#> [[23]]$currency_code
#> [1] "IDR"
#> 
#> [[23]]$unit
#> [1] 100
#> 
#> [[23]]$rate
#> [[23]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[23]]$rate$buying_rate
#> [1] 0.0302
#> 
#> [[23]]$rate$selling_rate
#> [1] 0.0302
#> 
#> [[23]]$rate$middle_rate
#> [1] 0.0302
#> 
#> 
#> 
#> [[24]]
#> [[24]]$currency_code
#> [1] "HKD"
#> 
#> [[24]]$unit
#> [1] 100
#> 
#> [[24]]$rate
#> [[24]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[24]]$rate$buying_rate
#> [1] 59.9979
#> 
#> [[24]]$rate$selling_rate
#> [1] 60.0387
#> 
#> [[24]]$rate$middle_rate
#> [1] 60.0183
#> 
#> 
#> 
#> [[25]]
#> [[25]]$currency_code
#> [1] "GBP"
#> 
#> [[25]]$unit
#> [1] 1
#> 
#> [[25]]$rate
#> [[25]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[25]]$rate$buying_rate
#> [1] 5.7926
#> 
#> [[25]]$rate$selling_rate
#> [1] 5.7982
#> 
#> [[25]]$rate$middle_rate
#> [1] 5.7954
#> 
#> 
#> 
#> [[26]]
#> [[26]]$currency_code
#> [1] "EUR"
#> 
#> [[26]]$unit
#> [1] 1
#> 
#> [[26]]$rate
#> [[26]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[26]]$rate$buying_rate
#> [1] 5.0748
#> 
#> [[26]]$rate$selling_rate
#> [1] 5.08
#> 
#> [[26]]$rate$middle_rate
#> [1] 5.0774
#> 
#> 
#> 
#> [[27]]
#> [[27]]$currency_code
#> [1] "EGP"
#> 
#> [[27]]$unit
#> [1] 1
#> 
#> [[27]]$rate
#> [[27]]$rate$date
#> [1] "2023-11-17"
#> 
#> [[27]]$rate$buying_rate
#> [1] 0.1512
#> 
#> [[27]]$rate$selling_rate
#> [1] 0.1515
#> 
#> [[27]]$rate$middle_rate
#> [1] 0.1514
#> 
#> 
#>