Interbank money market/deposit rates and volume

Obtain daily interbank money market rates and volumes of transactions according to tenure (2015 - present) from the BNM API.

Usage

get_islamic_interbank_rate(date = NULL, year = NULL, month = NULL)

get_interest_volume(
  product = "money_market_operations",
  date = NULL,
  year = NULL,
  month = NULL
)

get_interest_rate(
  product = "money_market_operations",
  date = NULL,
  year = NULL,
  month = NULL
)

Source

https://apikijangportal.bnm.gov.my/

Arguments

date

Character string of date with format as defined by RFC 3339, section 5.6 (YYYY-MM-DD). If specified, return values for the specified date.

year, month

Year and month as integers. If date, year and month left blank, return today's values.

product

One of "money_market_operations", "interbank" or "overall"

Details

Interbank money market rate: Daily interbank money market rates and volumes of transactions according to tenure. (2015 - present) Interbank money market volume: Daily interbank money market rates and volumes of transactions according to tenure. (2015 - present) Islamic interbank deposit rate: Daily weighted average of Islamic interbank deposit rates for various tenures. (Jan 2015-present)

Functions

get_islamic_interbank_rate(): Islamic interbank deposit rate

get_interest_volume(): Interbank money market volume

get_interest_rate(): Interbank money market rate

Examples

if (FALSE) get_interest_rate() # \dontrun{}
get_interest_rate(date = "2018-01-01")
#> $date
#> [1] "2018-01-01"
#> 
#> $overnight
#> [1] 2.91
#> 
#> $`1_week`
#> [1] 3.02
#> 
#> $`1_month`
#> [1] 3.08
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 
get_interest_rate(year = 2016, month = 2)
#> [[1]]
#> [[1]]$date
#> [1] "2016-02-01"
#> 
#> [[1]]$overnight
#> [1] 3.07
#> 
#> [[1]]$`1_week`
#> [1] 3.14
#> 
#> [[1]]$`1_month`
#> NULL
#> 
#> [[1]]$`3_month`
#> NULL
#> 
#> [[1]]$`6_month`
#> NULL
#> 
#> [[1]]$`1_year`
#> NULL
#> 
#> 
#> [[2]]
#> [[2]]$date
#> [1] "2016-02-02"
#> 
#> [[2]]$overnight
#> [1] 3.1
#> 
#> [[2]]$`1_week`
#> [1] 3.15
#> 
#> [[2]]$`1_month`
#> [1] 3.2
#> 
#> [[2]]$`3_month`
#> NULL
#> 
#> [[2]]$`6_month`
#> NULL
#> 
#> [[2]]$`1_year`
#> NULL
#> 
#> 
#> [[3]]
#> [[3]]$date
#> [1] "2016-02-03"
#> 
#> [[3]]$overnight
#> [1] 3.23
#> 
#> [[3]]$`1_week`
#> NULL
#> 
#> [[3]]$`1_month`
#> NULL
#> 
#> [[3]]$`3_month`
#> NULL
#> 
#> [[3]]$`6_month`
#> NULL
#> 
#> [[3]]$`1_year`
#> NULL
#> 
#> 
#> [[4]]
#> [[4]]$date
#> [1] "2016-02-04"
#> 
#> [[4]]$overnight
#> [1] 3.13
#> 
#> [[4]]$`1_week`
#> [1] 3.14
#> 
#> [[4]]$`1_month`
#> NULL
#> 
#> [[4]]$`3_month`
#> NULL
#> 
#> [[4]]$`6_month`
#> NULL
#> 
#> [[4]]$`1_year`
#> NULL
#> 
#> 
#> [[5]]
#> [[5]]$date
#> [1] "2016-02-05"
#> 
#> [[5]]$overnight
#> [1] 3.06
#> 
#> [[5]]$`1_week`
#> NULL
#> 
#> [[5]]$`1_month`
#> NULL
#> 
#> [[5]]$`3_month`
#> NULL
#> 
#> [[5]]$`6_month`
#> NULL
#> 
#> [[5]]$`1_year`
#> NULL
#> 
#> 
#> [[6]]
#> [[6]]$date
#> [1] "2016-02-06"
#> 
#> [[6]]$overnight
#> [1] 3.06
#> 
#> [[6]]$`1_week`
#> NULL
#> 
#> [[6]]$`1_month`
#> NULL
#> 
#> [[6]]$`3_month`
#> NULL
#> 
#> [[6]]$`6_month`
#> NULL
#> 
#> [[6]]$`1_year`
#> NULL
#> 
#> 
#> [[7]]
#> [[7]]$date
#> [1] "2016-02-07"
#> 
#> [[7]]$overnight
#> [1] 3.06
#> 
#> [[7]]$`1_week`
#> NULL
#> 
#> [[7]]$`1_month`
#> NULL
#> 
#> [[7]]$`3_month`
#> NULL
#> 
#> [[7]]$`6_month`
#> NULL
#> 
#> [[7]]$`1_year`
#> NULL
#> 
#> 
#> [[8]]
#> [[8]]$date
#> [1] "2016-02-08"
#> 
#> [[8]]$overnight
#> [1] 3.06
#> 
#> [[8]]$`1_week`
#> NULL
#> 
#> [[8]]$`1_month`
#> NULL
#> 
#> [[8]]$`3_month`
#> NULL
#> 
#> [[8]]$`6_month`
#> NULL
#> 
#> [[8]]$`1_year`
#> NULL
#> 
#> 
#> [[9]]
#> [[9]]$date
#> [1] "2016-02-09"
#> 
#> [[9]]$overnight
#> [1] 3.06
#> 
#> [[9]]$`1_week`
#> NULL
#> 
#> [[9]]$`1_month`
#> NULL
#> 
#> [[9]]$`3_month`
#> NULL
#> 
#> [[9]]$`6_month`
#> NULL
#> 
#> [[9]]$`1_year`
#> NULL
#> 
#> 
#> [[10]]
#> [[10]]$date
#> [1] "2016-02-10"
#> 
#> [[10]]$overnight
#> [1] 3.09
#> 
#> [[10]]$`1_week`
#> [1] 3.14
#> 
#> [[10]]$`1_month`
#> NULL
#> 
#> [[10]]$`3_month`
#> NULL
#> 
#> [[10]]$`6_month`
#> NULL
#> 
#> [[10]]$`1_year`
#> NULL
#> 
#> 
#> [[11]]
#> [[11]]$date
#> [1] "2016-02-11"
#> 
#> [[11]]$overnight
#> [1] 3.05
#> 
#> [[11]]$`1_week`
#> [1] 3.12
#> 
#> [[11]]$`1_month`
#> NULL
#> 
#> [[11]]$`3_month`
#> NULL
#> 
#> [[11]]$`6_month`
#> NULL
#> 
#> [[11]]$`1_year`
#> NULL
#> 
#> 
#> [[12]]
#> [[12]]$date
#> [1] "2016-02-12"
#> 
#> [[12]]$overnight
#> [1] 3.11
#> 
#> [[12]]$`1_week`
#> [1] 3.13
#> 
#> [[12]]$`1_month`
#> [1] 3.2
#> 
#> [[12]]$`3_month`
#> NULL
#> 
#> [[12]]$`6_month`
#> NULL
#> 
#> [[12]]$`1_year`
#> NULL
#> 
#> 
#> [[13]]
#> [[13]]$date
#> [1] "2016-02-13"
#> 
#> [[13]]$overnight
#> [1] 3.11
#> 
#> [[13]]$`1_week`
#> [1] 3.13
#> 
#> [[13]]$`1_month`
#> [1] 3.2
#> 
#> [[13]]$`3_month`
#> NULL
#> 
#> [[13]]$`6_month`
#> NULL
#> 
#> [[13]]$`1_year`
#> NULL
#> 
#> 
#> [[14]]
#> [[14]]$date
#> [1] "2016-02-14"
#> 
#> [[14]]$overnight
#> [1] 3.11
#> 
#> [[14]]$`1_week`
#> [1] 3.13
#> 
#> [[14]]$`1_month`
#> [1] 3.2
#> 
#> [[14]]$`3_month`
#> NULL
#> 
#> [[14]]$`6_month`
#> NULL
#> 
#> [[14]]$`1_year`
#> NULL
#> 
#> 
#> [[15]]
#> [[15]]$date
#> [1] "2016-02-15"
#> 
#> [[15]]$overnight
#> [1] 3.04
#> 
#> [[15]]$`1_week`
#> [1] 3.12
#> 
#> [[15]]$`1_month`
#> NULL
#> 
#> [[15]]$`3_month`
#> NULL
#> 
#> [[15]]$`6_month`
#> NULL
#> 
#> [[15]]$`1_year`
#> NULL
#> 
#> 
#> [[16]]
#> [[16]]$date
#> [1] "2016-02-16"
#> 
#> [[16]]$overnight
#> [1] 3.14
#> 
#> [[16]]$`1_week`
#> [1] 3.11
#> 
#> [[16]]$`1_month`
#> NULL
#> 
#> [[16]]$`3_month`
#> NULL
#> 
#> [[16]]$`6_month`
#> NULL
#> 
#> [[16]]$`1_year`
#> NULL
#> 
#> 
#> [[17]]
#> [[17]]$date
#> [1] "2016-02-17"
#> 
#> [[17]]$overnight
#> [1] 3.15
#> 
#> [[17]]$`1_week`
#> [1] 3.11
#> 
#> [[17]]$`1_month`
#> NULL
#> 
#> [[17]]$`3_month`
#> NULL
#> 
#> [[17]]$`6_month`
#> NULL
#> 
#> [[17]]$`1_year`
#> NULL
#> 
#> 
#> [[18]]
#> [[18]]$date
#> [1] "2016-02-18"
#> 
#> [[18]]$overnight
#> [1] 3.2
#> 
#> [[18]]$`1_week`
#> [1] 3.11
#> 
#> [[18]]$`1_month`
#> NULL
#> 
#> [[18]]$`3_month`
#> NULL
#> 
#> [[18]]$`6_month`
#> NULL
#> 
#> [[18]]$`1_year`
#> NULL
#> 
#> 
#> [[19]]
#> [[19]]$date
#> [1] "2016-02-19"
#> 
#> [[19]]$overnight
#> [1] 3.16
#> 
#> [[19]]$`1_week`
#> [1] 3.1
#> 
#> [[19]]$`1_month`
#> [1] 3.19
#> 
#> [[19]]$`3_month`
#> NULL
#> 
#> [[19]]$`6_month`
#> NULL
#> 
#> [[19]]$`1_year`
#> NULL
#> 
#> 
#> [[20]]
#> [[20]]$date
#> [1] "2016-02-20"
#> 
#> [[20]]$overnight
#> [1] 3.16
#> 
#> [[20]]$`1_week`
#> [1] 3.1
#> 
#> [[20]]$`1_month`
#> [1] 3.19
#> 
#> [[20]]$`3_month`
#> NULL
#> 
#> [[20]]$`6_month`
#> NULL
#> 
#> [[20]]$`1_year`
#> NULL
#> 
#> 
#> [[21]]
#> [[21]]$date
#> [1] "2016-02-21"
#> 
#> [[21]]$overnight
#> [1] 3.16
#> 
#> [[21]]$`1_week`
#> [1] 3.1
#> 
#> [[21]]$`1_month`
#> [1] 3.19
#> 
#> [[21]]$`3_month`
#> NULL
#> 
#> [[21]]$`6_month`
#> NULL
#> 
#> [[21]]$`1_year`
#> NULL
#> 
#> 
#> [[22]]
#> [[22]]$date
#> [1] "2016-02-22"
#> 
#> [[22]]$overnight
#> [1] 3.16
#> 
#> [[22]]$`1_week`
#> [1] 3.12
#> 
#> [[22]]$`1_month`
#> NULL
#> 
#> [[22]]$`3_month`
#> NULL
#> 
#> [[22]]$`6_month`
#> NULL
#> 
#> [[22]]$`1_year`
#> NULL
#> 
#> 
#> [[23]]
#> [[23]]$date
#> [1] "2016-02-23"
#> 
#> [[23]]$overnight
#> [1] 3.18
#> 
#> [[23]]$`1_week`
#> [1] 3.11
#> 
#> [[23]]$`1_month`
#> NULL
#> 
#> [[23]]$`3_month`
#> NULL
#> 
#> [[23]]$`6_month`
#> NULL
#> 
#> [[23]]$`1_year`
#> NULL
#> 
#> 
#> [[24]]
#> [[24]]$date
#> [1] "2016-02-24"
#> 
#> [[24]]$overnight
#> [1] 3.19
#> 
#> [[24]]$`1_week`
#> [1] 3.13
#> 
#> [[24]]$`1_month`
#> NULL
#> 
#> [[24]]$`3_month`
#> NULL
#> 
#> [[24]]$`6_month`
#> NULL
#> 
#> [[24]]$`1_year`
#> NULL
#> 
#> 
#> [[25]]
#> [[25]]$date
#> [1] "2016-02-25"
#> 
#> [[25]]$overnight
#> [1] 3.1
#> 
#> [[25]]$`1_week`
#> [1] 3.13
#> 
#> [[25]]$`1_month`
#> NULL
#> 
#> [[25]]$`3_month`
#> NULL
#> 
#> [[25]]$`6_month`
#> NULL
#> 
#> [[25]]$`1_year`
#> NULL
#> 
#> 
#> [[26]]
#> [[26]]$date
#> [1] "2016-02-26"
#> 
#> [[26]]$overnight
#> [1] 3.24
#> 
#> [[26]]$`1_week`
#> [1] 3.14
#> 
#> [[26]]$`1_month`
#> NULL
#> 
#> [[26]]$`3_month`
#> NULL
#> 
#> [[26]]$`6_month`
#> NULL
#> 
#> [[26]]$`1_year`
#> NULL
#> 
#> 
#> [[27]]
#> [[27]]$date
#> [1] "2016-02-27"
#> 
#> [[27]]$overnight
#> [1] 3.24
#> 
#> [[27]]$`1_week`
#> [1] 3.14
#> 
#> [[27]]$`1_month`
#> NULL
#> 
#> [[27]]$`3_month`
#> NULL
#> 
#> [[27]]$`6_month`
#> NULL
#> 
#> [[27]]$`1_year`
#> NULL
#> 
#> 
#> [[28]]
#> [[28]]$date
#> [1] "2016-02-28"
#> 
#> [[28]]$overnight
#> [1] 3.24
#> 
#> [[28]]$`1_week`
#> [1] 3.14
#> 
#> [[28]]$`1_month`
#> NULL
#> 
#> [[28]]$`3_month`
#> NULL
#> 
#> [[28]]$`6_month`
#> NULL
#> 
#> [[28]]$`1_year`
#> NULL
#> 
#> 
#> [[29]]
#> [[29]]$date
#> [1] "2016-02-29"
#> 
#> [[29]]$overnight
#> [1] 3.19
#> 
#> [[29]]$`1_week`
#> [1] 3.15
#> 
#> [[29]]$`1_month`
#> NULL
#> 
#> [[29]]$`3_month`
#> NULL
#> 
#> [[29]]$`6_month`
#> NULL
#> 
#> [[29]]$`1_year`
#> NULL
#> 
#> 
get_interest_rate(product = "overall", year = 2016, month = 2)
#> [[1]]
#> [[1]]$date
#> [1] "2016-02-01"
#> 
#> [[1]]$overnight
#> [1] 3.07
#> 
#> [[1]]$`1_week`
#> [1] 3.14
#> 
#> [[1]]$`1_month`
#> [1] 3.5
#> 
#> [[1]]$`3_month`
#> NULL
#> 
#> [[1]]$`6_month`
#> NULL
#> 
#> [[1]]$`1_year`
#> NULL
#> 
#> 
#> [[2]]
#> [[2]]$date
#> [1] "2016-02-02"
#> 
#> [[2]]$overnight
#> [1] 3.1
#> 
#> [[2]]$`1_week`
#> [1] 3.16
#> 
#> [[2]]$`1_month`
#> [1] 3.2
#> 
#> [[2]]$`3_month`
#> NULL
#> 
#> [[2]]$`6_month`
#> NULL
#> 
#> [[2]]$`1_year`
#> NULL
#> 
#> 
#> [[3]]
#> [[3]]$date
#> [1] "2016-02-03"
#> 
#> [[3]]$overnight
#> [1] 3.22
#> 
#> [[3]]$`1_week`
#> [1] 3.24
#> 
#> [[3]]$`1_month`
#> [1] 3.36
#> 
#> [[3]]$`3_month`
#> NULL
#> 
#> [[3]]$`6_month`
#> NULL
#> 
#> [[3]]$`1_year`
#> NULL
#> 
#> 
#> [[4]]
#> [[4]]$date
#> [1] "2016-02-04"
#> 
#> [[4]]$overnight
#> [1] 3.13
#> 
#> [[4]]$`1_week`
#> [1] 3.14
#> 
#> [[4]]$`1_month`
#> NULL
#> 
#> [[4]]$`3_month`
#> NULL
#> 
#> [[4]]$`6_month`
#> NULL
#> 
#> [[4]]$`1_year`
#> NULL
#> 
#> 
#> [[5]]
#> [[5]]$date
#> [1] "2016-02-05"
#> 
#> [[5]]$overnight
#> [1] 3.06
#> 
#> [[5]]$`1_week`
#> [1] 3.22
#> 
#> [[5]]$`1_month`
#> NULL
#> 
#> [[5]]$`3_month`
#> NULL
#> 
#> [[5]]$`6_month`
#> NULL
#> 
#> [[5]]$`1_year`
#> NULL
#> 
#> 
#> [[6]]
#> [[6]]$date
#> [1] "2016-02-06"
#> 
#> [[6]]$overnight
#> [1] 3.06
#> 
#> [[6]]$`1_week`
#> [1] 3.22
#> 
#> [[6]]$`1_month`
#> NULL
#> 
#> [[6]]$`3_month`
#> NULL
#> 
#> [[6]]$`6_month`
#> NULL
#> 
#> [[6]]$`1_year`
#> NULL
#> 
#> 
#> [[7]]
#> [[7]]$date
#> [1] "2016-02-07"
#> 
#> [[7]]$overnight
#> [1] 3.06
#> 
#> [[7]]$`1_week`
#> [1] 3.22
#> 
#> [[7]]$`1_month`
#> NULL
#> 
#> [[7]]$`3_month`
#> NULL
#> 
#> [[7]]$`6_month`
#> NULL
#> 
#> [[7]]$`1_year`
#> NULL
#> 
#> 
#> [[8]]
#> [[8]]$date
#> [1] "2016-02-08"
#> 
#> [[8]]$overnight
#> [1] 3.06
#> 
#> [[8]]$`1_week`
#> [1] 3.22
#> 
#> [[8]]$`1_month`
#> NULL
#> 
#> [[8]]$`3_month`
#> NULL
#> 
#> [[8]]$`6_month`
#> NULL
#> 
#> [[8]]$`1_year`
#> NULL
#> 
#> 
#> [[9]]
#> [[9]]$date
#> [1] "2016-02-10"
#> 
#> [[9]]$overnight
#> [1] 3.09
#> 
#> [[9]]$`1_week`
#> [1] 3.14
#> 
#> [[9]]$`1_month`
#> NULL
#> 
#> [[9]]$`3_month`
#> NULL
#> 
#> [[9]]$`6_month`
#> NULL
#> 
#> [[9]]$`1_year`
#> NULL
#> 
#> 
#> [[10]]
#> [[10]]$date
#> [1] "2016-02-11"
#> 
#> [[10]]$overnight
#> [1] 3.05
#> 
#> [[10]]$`1_week`
#> [1] 3.16
#> 
#> [[10]]$`1_month`
#> NULL
#> 
#> [[10]]$`3_month`
#> NULL
#> 
#> [[10]]$`6_month`
#> NULL
#> 
#> [[10]]$`1_year`
#> NULL
#> 
#> 
#> [[11]]
#> [[11]]$date
#> [1] "2016-02-12"
#> 
#> [[11]]$overnight
#> [1] 3.11
#> 
#> [[11]]$`1_week`
#> [1] 3.13
#> 
#> [[11]]$`1_month`
#> [1] 3.2
#> 
#> [[11]]$`3_month`
#> NULL
#> 
#> [[11]]$`6_month`
#> NULL
#> 
#> [[11]]$`1_year`
#> NULL
#> 
#> 
#> [[12]]
#> [[12]]$date
#> [1] "2016-02-13"
#> 
#> [[12]]$overnight
#> [1] 3.11
#> 
#> [[12]]$`1_week`
#> [1] 3.13
#> 
#> [[12]]$`1_month`
#> [1] 3.2
#> 
#> [[12]]$`3_month`
#> NULL
#> 
#> [[12]]$`6_month`
#> NULL
#> 
#> [[12]]$`1_year`
#> NULL
#> 
#> 
#> [[13]]
#> [[13]]$date
#> [1] "2016-02-14"
#> 
#> [[13]]$overnight
#> [1] 3.11
#> 
#> [[13]]$`1_week`
#> [1] 3.13
#> 
#> [[13]]$`1_month`
#> [1] 3.2
#> 
#> [[13]]$`3_month`
#> NULL
#> 
#> [[13]]$`6_month`
#> NULL
#> 
#> [[13]]$`1_year`
#> NULL
#> 
#> 
#> [[14]]
#> [[14]]$date
#> [1] "2016-02-15"
#> 
#> [[14]]$overnight
#> [1] 3.04
#> 
#> [[14]]$`1_week`
#> [1] 3.12
#> 
#> [[14]]$`1_month`
#> [1] 3.36
#> 
#> [[14]]$`3_month`
#> NULL
#> 
#> [[14]]$`6_month`
#> NULL
#> 
#> [[14]]$`1_year`
#> NULL
#> 
#> 
#> [[15]]
#> [[15]]$date
#> [1] "2016-02-16"
#> 
#> [[15]]$overnight
#> [1] 3.14
#> 
#> [[15]]$`1_week`
#> [1] 3.13
#> 
#> [[15]]$`1_month`
#> [1] 3.1
#> 
#> [[15]]$`3_month`
#> NULL
#> 
#> [[15]]$`6_month`
#> NULL
#> 
#> [[15]]$`1_year`
#> NULL
#> 
#> 
#> [[16]]
#> [[16]]$date
#> [1] "2016-02-17"
#> 
#> [[16]]$overnight
#> [1] 3.14
#> 
#> [[16]]$`1_week`
#> [1] 3.11
#> 
#> [[16]]$`1_month`
#> NULL
#> 
#> [[16]]$`3_month`
#> NULL
#> 
#> [[16]]$`6_month`
#> NULL
#> 
#> [[16]]$`1_year`
#> NULL
#> 
#> 
#> [[17]]
#> [[17]]$date
#> [1] "2016-02-18"
#> 
#> [[17]]$overnight
#> [1] 3.2
#> 
#> [[17]]$`1_week`
#> [1] 3.11
#> 
#> [[17]]$`1_month`
#> [1] 3.37
#> 
#> [[17]]$`3_month`
#> NULL
#> 
#> [[17]]$`6_month`
#> NULL
#> 
#> [[17]]$`1_year`
#> NULL
#> 
#> 
#> [[18]]
#> [[18]]$date
#> [1] "2016-02-19"
#> 
#> [[18]]$overnight
#> [1] 3.16
#> 
#> [[18]]$`1_week`
#> [1] 3.11
#> 
#> [[18]]$`1_month`
#> [1] 3.19
#> 
#> [[18]]$`3_month`
#> NULL
#> 
#> [[18]]$`6_month`
#> NULL
#> 
#> [[18]]$`1_year`
#> NULL
#> 
#> 
#> [[19]]
#> [[19]]$date
#> [1] "2016-02-20"
#> 
#> [[19]]$overnight
#> [1] 3.16
#> 
#> [[19]]$`1_week`
#> [1] 3.11
#> 
#> [[19]]$`1_month`
#> [1] 3.19
#> 
#> [[19]]$`3_month`
#> NULL
#> 
#> [[19]]$`6_month`
#> NULL
#> 
#> [[19]]$`1_year`
#> NULL
#> 
#> 
#> [[20]]
#> [[20]]$date
#> [1] "2016-02-21"
#> 
#> [[20]]$overnight
#> [1] 3.16
#> 
#> [[20]]$`1_week`
#> [1] 3.11
#> 
#> [[20]]$`1_month`
#> [1] 3.19
#> 
#> [[20]]$`3_month`
#> NULL
#> 
#> [[20]]$`6_month`
#> NULL
#> 
#> [[20]]$`1_year`
#> NULL
#> 
#> 
#> [[21]]
#> [[21]]$date
#> [1] "2016-02-22"
#> 
#> [[21]]$overnight
#> [1] 3.15
#> 
#> [[21]]$`1_week`
#> [1] 3.12
#> 
#> [[21]]$`1_month`
#> NULL
#> 
#> [[21]]$`3_month`
#> NULL
#> 
#> [[21]]$`6_month`
#> NULL
#> 
#> [[21]]$`1_year`
#> NULL
#> 
#> 
#> [[22]]
#> [[22]]$date
#> [1] "2016-02-23"
#> 
#> [[22]]$overnight
#> [1] 3.18
#> 
#> [[22]]$`1_week`
#> [1] 3.23
#> 
#> [[22]]$`1_month`
#> NULL
#> 
#> [[22]]$`3_month`
#> NULL
#> 
#> [[22]]$`6_month`
#> NULL
#> 
#> [[22]]$`1_year`
#> NULL
#> 
#> 
#> [[23]]
#> [[23]]$date
#> [1] "2016-02-24"
#> 
#> [[23]]$overnight
#> [1] 3.19
#> 
#> [[23]]$`1_week`
#> [1] 3.13
#> 
#> [[23]]$`1_month`
#> NULL
#> 
#> [[23]]$`3_month`
#> NULL
#> 
#> [[23]]$`6_month`
#> NULL
#> 
#> [[23]]$`1_year`
#> NULL
#> 
#> 
#> [[24]]
#> [[24]]$date
#> [1] "2016-02-25"
#> 
#> [[24]]$overnight
#> [1] 3.1
#> 
#> [[24]]$`1_week`
#> [1] 3.13
#> 
#> [[24]]$`1_month`
#> [1] 3.37
#> 
#> [[24]]$`3_month`
#> NULL
#> 
#> [[24]]$`6_month`
#> NULL
#> 
#> [[24]]$`1_year`
#> NULL
#> 
#> 
#> [[25]]
#> [[25]]$date
#> [1] "2016-02-26"
#> 
#> [[25]]$overnight
#> [1] 3.23
#> 
#> [[25]]$`1_week`
#> [1] 3.14
#> 
#> [[25]]$`1_month`
#> [1] 3.38
#> 
#> [[25]]$`3_month`
#> NULL
#> 
#> [[25]]$`6_month`
#> NULL
#> 
#> [[25]]$`1_year`
#> NULL
#> 
#> 
#> [[26]]
#> [[26]]$date
#> [1] "2016-02-27"
#> 
#> [[26]]$overnight
#> [1] 3.23
#> 
#> [[26]]$`1_week`
#> [1] 3.14
#> 
#> [[26]]$`1_month`
#> [1] 3.38
#> 
#> [[26]]$`3_month`
#> NULL
#> 
#> [[26]]$`6_month`
#> NULL
#> 
#> [[26]]$`1_year`
#> NULL
#> 
#> 
#> [[27]]
#> [[27]]$date
#> [1] "2016-02-28"
#> 
#> [[27]]$overnight
#> [1] 3.23
#> 
#> [[27]]$`1_week`
#> [1] 3.14
#> 
#> [[27]]$`1_month`
#> [1] 3.38
#> 
#> [[27]]$`3_month`
#> NULL
#> 
#> [[27]]$`6_month`
#> NULL
#> 
#> [[27]]$`1_year`
#> NULL
#> 
#> 
#> [[28]]
#> [[28]]$date
#> [1] "2016-02-29"
#> 
#> [[28]]$overnight
#> [1] 3.19
#> 
#> [[28]]$`1_week`
#> [1] 3.16
#> 
#> [[28]]$`1_month`
#> [1] 3.41
#> 
#> [[28]]$`3_month`
#> NULL
#> 
#> [[28]]$`6_month`
#> NULL
#> 
#> [[28]]$`1_year`
#> NULL
#> 
#> 
if (FALSE) get_interest_volume() # \dontrun{}
get_interest_volume(date = "2018-01-01")
#> $date
#> [1] "2018-01-01"
#> 
#> $overnight
#> [1] 33100
#> 
#> $`1_week`
#> [1] 1430
#> 
#> $`1_month`
#> [1] 200
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 
#> $other
#> [1] 5030
#> 
get_interest_volume(year = 2016, month = 2)
#> [[1]]
#> [[1]]$date
#> [1] "2016-02-01"
#> 
#> [[1]]$overnight
#> [1] 32000
#> 
#> [[1]]$`1_week`
#> [1] 3860
#> 
#> [[1]]$`1_month`
#> NULL
#> 
#> [[1]]$`3_month`
#> NULL
#> 
#> [[1]]$`6_month`
#> NULL
#> 
#> [[1]]$`1_year`
#> NULL
#> 
#> [[1]]$other
#> [1] 3140
#> 
#> 
#> [[2]]
#> [[2]]$date
#> [1] "2016-02-02"
#> 
#> [[2]]$overnight
#> [1] 32000
#> 
#> [[2]]$`1_week`
#> [1] 1860
#> 
#> [[2]]$`1_month`
#> [1] 7030
#> 
#> [[2]]$`3_month`
#> NULL
#> 
#> [[2]]$`6_month`
#> NULL
#> 
#> [[2]]$`1_year`
#> NULL
#> 
#> [[2]]$other
#> [1] 4355
#> 
#> 
#> [[3]]
#> [[3]]$date
#> [1] "2016-02-03"
#> 
#> [[3]]$overnight
#> [1] 32000
#> 
#> [[3]]$`1_week`
#> NULL
#> 
#> [[3]]$`1_month`
#> NULL
#> 
#> [[3]]$`3_month`
#> NULL
#> 
#> [[3]]$`6_month`
#> NULL
#> 
#> [[3]]$`1_year`
#> NULL
#> 
#> [[3]]$other
#> [1] 4180
#> 
#> 
#> [[4]]
#> [[4]]$date
#> [1] "2016-02-04"
#> 
#> [[4]]$overnight
#> [1] 32000
#> 
#> [[4]]$`1_week`
#> [1] 3000
#> 
#> [[4]]$`1_month`
#> NULL
#> 
#> [[4]]$`3_month`
#> NULL
#> 
#> [[4]]$`6_month`
#> NULL
#> 
#> [[4]]$`1_year`
#> NULL
#> 
#> [[4]]$other
#> NULL
#> 
#> 
#> [[5]]
#> [[5]]$date
#> [1] "2016-02-05"
#> 
#> [[5]]$overnight
#> [1] 30000
#> 
#> [[5]]$`1_week`
#> NULL
#> 
#> [[5]]$`1_month`
#> NULL
#> 
#> [[5]]$`3_month`
#> NULL
#> 
#> [[5]]$`6_month`
#> NULL
#> 
#> [[5]]$`1_year`
#> NULL
#> 
#> [[5]]$other
#> [1] 3050
#> 
#> 
#> [[6]]
#> [[6]]$date
#> [1] "2016-02-06"
#> 
#> [[6]]$overnight
#> [1] 30000
#> 
#> [[6]]$`1_week`
#> NULL
#> 
#> [[6]]$`1_month`
#> NULL
#> 
#> [[6]]$`3_month`
#> NULL
#> 
#> [[6]]$`6_month`
#> NULL
#> 
#> [[6]]$`1_year`
#> NULL
#> 
#> [[6]]$other
#> [1] 3050
#> 
#> 
#> [[7]]
#> [[7]]$date
#> [1] "2016-02-07"
#> 
#> [[7]]$overnight
#> [1] 30000
#> 
#> [[7]]$`1_week`
#> NULL
#> 
#> [[7]]$`1_month`
#> NULL
#> 
#> [[7]]$`3_month`
#> NULL
#> 
#> [[7]]$`6_month`
#> NULL
#> 
#> [[7]]$`1_year`
#> NULL
#> 
#> [[7]]$other
#> [1] 3050
#> 
#> 
#> [[8]]
#> [[8]]$date
#> [1] "2016-02-08"
#> 
#> [[8]]$overnight
#> [1] 30000
#> 
#> [[8]]$`1_week`
#> NULL
#> 
#> [[8]]$`1_month`
#> NULL
#> 
#> [[8]]$`3_month`
#> NULL
#> 
#> [[8]]$`6_month`
#> NULL
#> 
#> [[8]]$`1_year`
#> NULL
#> 
#> [[8]]$other
#> [1] 3050
#> 
#> 
#> [[9]]
#> [[9]]$date
#> [1] "2016-02-09"
#> 
#> [[9]]$overnight
#> [1] 30000
#> 
#> [[9]]$`1_week`
#> NULL
#> 
#> [[9]]$`1_month`
#> NULL
#> 
#> [[9]]$`3_month`
#> NULL
#> 
#> [[9]]$`6_month`
#> NULL
#> 
#> [[9]]$`1_year`
#> NULL
#> 
#> [[9]]$other
#> [1] 3050
#> 
#> 
#> [[10]]
#> [[10]]$date
#> [1] "2016-02-10"
#> 
#> [[10]]$overnight
#> [1] 30000
#> 
#> [[10]]$`1_week`
#> [1] 3800
#> 
#> [[10]]$`1_month`
#> NULL
#> 
#> [[10]]$`3_month`
#> NULL
#> 
#> [[10]]$`6_month`
#> NULL
#> 
#> [[10]]$`1_year`
#> NULL
#> 
#> [[10]]$other
#> [1] 3280
#> 
#> 
#> [[11]]
#> [[11]]$date
#> [1] "2016-02-11"
#> 
#> [[11]]$overnight
#> [1] 32000
#> 
#> [[11]]$`1_week`
#> [1] 1500
#> 
#> [[11]]$`1_month`
#> NULL
#> 
#> [[11]]$`3_month`
#> NULL
#> 
#> [[11]]$`6_month`
#> NULL
#> 
#> [[11]]$`1_year`
#> NULL
#> 
#> [[11]]$other
#> [1] 1000
#> 
#> 
#> [[12]]
#> [[12]]$date
#> [1] "2016-02-12"
#> 
#> [[12]]$overnight
#> [1] 31000
#> 
#> [[12]]$`1_week`
#> [1] 170
#> 
#> [[12]]$`1_month`
#> [1] 4165
#> 
#> [[12]]$`3_month`
#> NULL
#> 
#> [[12]]$`6_month`
#> NULL
#> 
#> [[12]]$`1_year`
#> NULL
#> 
#> [[12]]$other
#> [1] 2030
#> 
#> 
#> [[13]]
#> [[13]]$date
#> [1] "2016-02-13"
#> 
#> [[13]]$overnight
#> [1] 31000
#> 
#> [[13]]$`1_week`
#> [1] 170
#> 
#> [[13]]$`1_month`
#> [1] 4165
#> 
#> [[13]]$`3_month`
#> NULL
#> 
#> [[13]]$`6_month`
#> NULL
#> 
#> [[13]]$`1_year`
#> NULL
#> 
#> [[13]]$other
#> [1] 2030
#> 
#> 
#> [[14]]
#> [[14]]$date
#> [1] "2016-02-14"
#> 
#> [[14]]$overnight
#> [1] 31000
#> 
#> [[14]]$`1_week`
#> [1] 170
#> 
#> [[14]]$`1_month`
#> [1] 4165
#> 
#> [[14]]$`3_month`
#> NULL
#> 
#> [[14]]$`6_month`
#> NULL
#> 
#> [[14]]$`1_year`
#> NULL
#> 
#> [[14]]$other
#> [1] 2030
#> 
#> 
#> [[15]]
#> [[15]]$date
#> [1] "2016-02-15"
#> 
#> [[15]]$overnight
#> [1] 30000
#> 
#> [[15]]$`1_week`
#> [1] 2000
#> 
#> [[15]]$`1_month`
#> NULL
#> 
#> [[15]]$`3_month`
#> NULL
#> 
#> [[15]]$`6_month`
#> NULL
#> 
#> [[15]]$`1_year`
#> NULL
#> 
#> [[15]]$other
#> NULL
#> 
#> 
#> [[16]]
#> [[16]]$date
#> [1] "2016-02-16"
#> 
#> [[16]]$overnight
#> [1] 32000
#> 
#> [[16]]$`1_week`
#> [1] 2000
#> 
#> [[16]]$`1_month`
#> NULL
#> 
#> [[16]]$`3_month`
#> NULL
#> 
#> [[16]]$`6_month`
#> NULL
#> 
#> [[16]]$`1_year`
#> NULL
#> 
#> [[16]]$other
#> NULL
#> 
#> 
#> [[17]]
#> [[17]]$date
#> [1] "2016-02-17"
#> 
#> [[17]]$overnight
#> [1] 31000
#> 
#> [[17]]$`1_week`
#> [1] 2000
#> 
#> [[17]]$`1_month`
#> NULL
#> 
#> [[17]]$`3_month`
#> NULL
#> 
#> [[17]]$`6_month`
#> NULL
#> 
#> [[17]]$`1_year`
#> NULL
#> 
#> [[17]]$other
#> NULL
#> 
#> 
#> [[18]]
#> [[18]]$date
#> [1] "2016-02-18"
#> 
#> [[18]]$overnight
#> [1] 33000
#> 
#> [[18]]$`1_week`
#> [1] 4000
#> 
#> [[18]]$`1_month`
#> NULL
#> 
#> [[18]]$`3_month`
#> NULL
#> 
#> [[18]]$`6_month`
#> NULL
#> 
#> [[18]]$`1_year`
#> NULL
#> 
#> [[18]]$other
#> NULL
#> 
#> 
#> [[19]]
#> [[19]]$date
#> [1] "2016-02-19"
#> 
#> [[19]]$overnight
#> [1] 31000
#> 
#> [[19]]$`1_week`
#> [1] 450
#> 
#> [[19]]$`1_month`
#> [1] 3270
#> 
#> [[19]]$`3_month`
#> NULL
#> 
#> [[19]]$`6_month`
#> NULL
#> 
#> [[19]]$`1_year`
#> NULL
#> 
#> [[19]]$other
#> [1] 3430
#> 
#> 
#> [[20]]
#> [[20]]$date
#> [1] "2016-02-20"
#> 
#> [[20]]$overnight
#> [1] 31000
#> 
#> [[20]]$`1_week`
#> [1] 450
#> 
#> [[20]]$`1_month`
#> [1] 3270
#> 
#> [[20]]$`3_month`
#> NULL
#> 
#> [[20]]$`6_month`
#> NULL
#> 
#> [[20]]$`1_year`
#> NULL
#> 
#> [[20]]$other
#> [1] 3430
#> 
#> 
#> [[21]]
#> [[21]]$date
#> [1] "2016-02-21"
#> 
#> [[21]]$overnight
#> [1] 31000
#> 
#> [[21]]$`1_week`
#> [1] 450
#> 
#> [[21]]$`1_month`
#> [1] 3270
#> 
#> [[21]]$`3_month`
#> NULL
#> 
#> [[21]]$`6_month`
#> NULL
#> 
#> [[21]]$`1_year`
#> NULL
#> 
#> [[21]]$other
#> [1] 3430
#> 
#> 
#> [[22]]
#> [[22]]$date
#> [1] "2016-02-22"
#> 
#> [[22]]$overnight
#> [1] 32000
#> 
#> [[22]]$`1_week`
#> [1] 4940
#> 
#> [[22]]$`1_month`
#> NULL
#> 
#> [[22]]$`3_month`
#> NULL
#> 
#> [[22]]$`6_month`
#> NULL
#> 
#> [[22]]$`1_year`
#> NULL
#> 
#> [[22]]$other
#> [1] 2000
#> 
#> 
#> [[23]]
#> [[23]]$date
#> [1] "2016-02-23"
#> 
#> [[23]]$overnight
#> [1] 33000
#> 
#> [[23]]$`1_week`
#> [1] 2880
#> 
#> [[23]]$`1_month`
#> NULL
#> 
#> [[23]]$`3_month`
#> NULL
#> 
#> [[23]]$`6_month`
#> NULL
#> 
#> [[23]]$`1_year`
#> NULL
#> 
#> [[23]]$other
#> [1] 2700
#> 
#> 
#> [[24]]
#> [[24]]$date
#> [1] "2016-02-24"
#> 
#> [[24]]$overnight
#> [1] 36000
#> 
#> [[24]]$`1_week`
#> [1] 3620
#> 
#> [[24]]$`1_month`
#> NULL
#> 
#> [[24]]$`3_month`
#> NULL
#> 
#> [[24]]$`6_month`
#> NULL
#> 
#> [[24]]$`1_year`
#> NULL
#> 
#> [[24]]$other
#> NULL
#> 
#> 
#> [[25]]
#> [[25]]$date
#> [1] "2016-02-25"
#> 
#> [[25]]$overnight
#> [1] 38000
#> 
#> [[25]]$`1_week`
#> [1] 3500
#> 
#> [[25]]$`1_month`
#> NULL
#> 
#> [[25]]$`3_month`
#> NULL
#> 
#> [[25]]$`6_month`
#> NULL
#> 
#> [[25]]$`1_year`
#> NULL
#> 
#> [[25]]$other
#> NULL
#> 
#> 
#> [[26]]
#> [[26]]$date
#> [1] "2016-02-26"
#> 
#> [[26]]$overnight
#> [1] 32692
#> 
#> [[26]]$`1_week`
#> [1] 4000
#> 
#> [[26]]$`1_month`
#> NULL
#> 
#> [[26]]$`3_month`
#> NULL
#> 
#> [[26]]$`6_month`
#> NULL
#> 
#> [[26]]$`1_year`
#> NULL
#> 
#> [[26]]$other
#> NULL
#> 
#> 
#> [[27]]
#> [[27]]$date
#> [1] "2016-02-27"
#> 
#> [[27]]$overnight
#> [1] 32692
#> 
#> [[27]]$`1_week`
#> [1] 4000
#> 
#> [[27]]$`1_month`
#> NULL
#> 
#> [[27]]$`3_month`
#> NULL
#> 
#> [[27]]$`6_month`
#> NULL
#> 
#> [[27]]$`1_year`
#> NULL
#> 
#> [[27]]$other
#> NULL
#> 
#> 
#> [[28]]
#> [[28]]$date
#> [1] "2016-02-28"
#> 
#> [[28]]$overnight
#> [1] 32692
#> 
#> [[28]]$`1_week`
#> [1] 4000
#> 
#> [[28]]$`1_month`
#> NULL
#> 
#> [[28]]$`3_month`
#> NULL
#> 
#> [[28]]$`6_month`
#> NULL
#> 
#> [[28]]$`1_year`
#> NULL
#> 
#> [[28]]$other
#> NULL
#> 
#> 
#> [[29]]
#> [[29]]$date
#> [1] "2016-02-29"
#> 
#> [[29]]$overnight
#> [1] 34000
#> 
#> [[29]]$`1_week`
#> [1] 4330
#> 
#> [[29]]$`1_month`
#> NULL
#> 
#> [[29]]$`3_month`
#> NULL
#> 
#> [[29]]$`6_month`
#> NULL
#> 
#> [[29]]$`1_year`
#> NULL
#> 
#> [[29]]$other
#> NULL
#> 
#> 
get_interest_volume(product = "overall", year = 2016, month = 2)
#> [[1]]
#> [[1]]$date
#> [1] "2016-02-01"
#> 
#> [[1]]$overnight
#> [1] 34068.13
#> 
#> [[1]]$`1_week`
#> [1] 3860
#> 
#> [[1]]$`1_month`
#> [1] 150
#> 
#> [[1]]$`3_month`
#> NULL
#> 
#> [[1]]$`6_month`
#> NULL
#> 
#> [[1]]$`1_year`
#> NULL
#> 
#> [[1]]$other
#> [1] 3362
#> 
#> 
#> [[2]]
#> [[2]]$date
#> [1] "2016-02-02"
#> 
#> [[2]]$overnight
#> [1] 33708.7
#> 
#> [[2]]$`1_week`
#> [1] 2160
#> 
#> [[2]]$`1_month`
#> [1] 7110
#> 
#> [[2]]$`3_month`
#> NULL
#> 
#> [[2]]$`6_month`
#> NULL
#> 
#> [[2]]$`1_year`
#> NULL
#> 
#> [[2]]$other
#> [1] 4355
#> 
#> 
#> [[3]]
#> [[3]]$date
#> [1] "2016-02-03"
#> 
#> [[3]]$overnight
#> [1] 34533.36
#> 
#> [[3]]$`1_week`
#> [1] 155
#> 
#> [[3]]$`1_month`
#> [1] 50
#> 
#> [[3]]$`3_month`
#> NULL
#> 
#> [[3]]$`6_month`
#> NULL
#> 
#> [[3]]$`1_year`
#> NULL
#> 
#> [[3]]$other
#> [1] 4180
#> 
#> 
#> [[4]]
#> [[4]]$date
#> [1] "2016-02-04"
#> 
#> [[4]]$overnight
#> [1] 34696.38
#> 
#> [[4]]$`1_week`
#> [1] 3290
#> 
#> [[4]]$`1_month`
#> NULL
#> 
#> [[4]]$`3_month`
#> NULL
#> 
#> [[4]]$`6_month`
#> NULL
#> 
#> [[4]]$`1_year`
#> NULL
#> 
#> [[4]]$other
#> [1] 150
#> 
#> 
#> [[5]]
#> [[5]]$date
#> [1] "2016-02-05"
#> 
#> [[5]]$overnight
#> [1] 32216.19
#> 
#> [[5]]$`1_week`
#> [1] 200
#> 
#> [[5]]$`1_month`
#> NULL
#> 
#> [[5]]$`3_month`
#> NULL
#> 
#> [[5]]$`6_month`
#> NULL
#> 
#> [[5]]$`1_year`
#> NULL
#> 
#> [[5]]$other
#> [1] 3750
#> 
#> 
#> [[6]]
#> [[6]]$date
#> [1] "2016-02-06"
#> 
#> [[6]]$overnight
#> [1] 32216.19
#> 
#> [[6]]$`1_week`
#> [1] 200
#> 
#> [[6]]$`1_month`
#> NULL
#> 
#> [[6]]$`3_month`
#> NULL
#> 
#> [[6]]$`6_month`
#> NULL
#> 
#> [[6]]$`1_year`
#> NULL
#> 
#> [[6]]$other
#> [1] 3750
#> 
#> 
#> [[7]]
#> [[7]]$date
#> [1] "2016-02-07"
#> 
#> [[7]]$overnight
#> [1] 32216.19
#> 
#> [[7]]$`1_week`
#> [1] 200
#> 
#> [[7]]$`1_month`
#> NULL
#> 
#> [[7]]$`3_month`
#> NULL
#> 
#> [[7]]$`6_month`
#> NULL
#> 
#> [[7]]$`1_year`
#> NULL
#> 
#> [[7]]$other
#> [1] 3750
#> 
#> 
#> [[8]]
#> [[8]]$date
#> [1] "2016-02-08"
#> 
#> [[8]]$overnight
#> [1] 32216.19
#> 
#> [[8]]$`1_week`
#> [1] 200
#> 
#> [[8]]$`1_month`
#> NULL
#> 
#> [[8]]$`3_month`
#> NULL
#> 
#> [[8]]$`6_month`
#> NULL
#> 
#> [[8]]$`1_year`
#> NULL
#> 
#> [[8]]$other
#> [1] 3750
#> 
#> 
#> [[9]]
#> [[9]]$date
#> [1] "2016-02-10"
#> 
#> [[9]]$overnight
#> [1] 32485.55
#> 
#> [[9]]$`1_week`
#> [1] 3900
#> 
#> [[9]]$`1_month`
#> NULL
#> 
#> [[9]]$`3_month`
#> NULL
#> 
#> [[9]]$`6_month`
#> NULL
#> 
#> [[9]]$`1_year`
#> NULL
#> 
#> [[9]]$other
#> [1] 3530
#> 
#> 
#> [[10]]
#> [[10]]$date
#> [1] "2016-02-11"
#> 
#> [[10]]$overnight
#> [1] 34464.27
#> 
#> [[10]]$`1_week`
#> [1] 2500
#> 
#> [[10]]$`1_month`
#> NULL
#> 
#> [[10]]$`3_month`
#> NULL
#> 
#> [[10]]$`6_month`
#> NULL
#> 
#> [[10]]$`1_year`
#> NULL
#> 
#> [[10]]$other
#> [1] 1250
#> 
#> 
#> [[11]]
#> [[11]]$date
#> [1] "2016-02-12"
#> 
#> [[11]]$overnight
#> [1] 32944.47
#> 
#> [[11]]$`1_week`
#> [1] 170
#> 
#> [[11]]$`1_month`
#> [1] 4165
#> 
#> [[11]]$`3_month`
#> NULL
#> 
#> [[11]]$`6_month`
#> NULL
#> 
#> [[11]]$`1_year`
#> NULL
#> 
#> [[11]]$other
#> [1] 2430
#> 
#> 
#> [[12]]
#> [[12]]$date
#> [1] "2016-02-13"
#> 
#> [[12]]$overnight
#> [1] 32944.47
#> 
#> [[12]]$`1_week`
#> [1] 170
#> 
#> [[12]]$`1_month`
#> [1] 4165
#> 
#> [[12]]$`3_month`
#> NULL
#> 
#> [[12]]$`6_month`
#> NULL
#> 
#> [[12]]$`1_year`
#> NULL
#> 
#> [[12]]$other
#> [1] 2430
#> 
#> 
#> [[13]]
#> [[13]]$date
#> [1] "2016-02-14"
#> 
#> [[13]]$overnight
#> [1] 32944.47
#> 
#> [[13]]$`1_week`
#> [1] 170
#> 
#> [[13]]$`1_month`
#> [1] 4165
#> 
#> [[13]]$`3_month`
#> NULL
#> 
#> [[13]]$`6_month`
#> NULL
#> 
#> [[13]]$`1_year`
#> NULL
#> 
#> [[13]]$other
#> [1] 2430
#> 
#> 
#> [[14]]
#> [[14]]$date
#> [1] "2016-02-15"
#> 
#> [[14]]$overnight
#> [1] 32324.53
#> 
#> [[14]]$`1_week`
#> [1] 2190
#> 
#> [[14]]$`1_month`
#> [1] 60
#> 
#> [[14]]$`3_month`
#> NULL
#> 
#> [[14]]$`6_month`
#> NULL
#> 
#> [[14]]$`1_year`
#> NULL
#> 
#> [[14]]$other
#> NULL
#> 
#> 
#> [[15]]
#> [[15]]$date
#> [1] "2016-02-16"
#> 
#> [[15]]$overnight
#> [1] 33780.16
#> 
#> [[15]]$`1_week`
#> [1] 2300
#> 
#> [[15]]$`1_month`
#> [1] 3.38
#> 
#> [[15]]$`3_month`
#> NULL
#> 
#> [[15]]$`6_month`
#> NULL
#> 
#> [[15]]$`1_year`
#> NULL
#> 
#> [[15]]$other
#> [1] 400
#> 
#> 
#> [[16]]
#> [[16]]$date
#> [1] "2016-02-17"
#> 
#> [[16]]$overnight
#> [1] 33542.95
#> 
#> [[16]]$`1_week`
#> [1] 2000
#> 
#> [[16]]$`1_month`
#> NULL
#> 
#> [[16]]$`3_month`
#> NULL
#> 
#> [[16]]$`6_month`
#> NULL
#> 
#> [[16]]$`1_year`
#> NULL
#> 
#> [[16]]$other
#> [1] 510
#> 
#> 
#> [[17]]
#> [[17]]$date
#> [1] "2016-02-18"
#> 
#> [[17]]$overnight
#> [1] 34370.64
#> 
#> [[17]]$`1_week`
#> [1] 4180
#> 
#> [[17]]$`1_month`
#> [1] 50
#> 
#> [[17]]$`3_month`
#> NULL
#> 
#> [[17]]$`6_month`
#> NULL
#> 
#> [[17]]$`1_year`
#> NULL
#> 
#> [[17]]$other
#> [1] 100
#> 
#> 
#> [[18]]
#> [[18]]$date
#> [1] "2016-02-19"
#> 
#> [[18]]$overnight
#> [1] 33369.02
#> 
#> [[18]]$`1_week`
#> [1] 500
#> 
#> [[18]]$`1_month`
#> [1] 3270
#> 
#> [[18]]$`3_month`
#> NULL
#> 
#> [[18]]$`6_month`
#> NULL
#> 
#> [[18]]$`1_year`
#> NULL
#> 
#> [[18]]$other
#> [1] 3530
#> 
#> 
#> [[19]]
#> [[19]]$date
#> [1] "2016-02-20"
#> 
#> [[19]]$overnight
#> [1] 33369.02
#> 
#> [[19]]$`1_week`
#> [1] 500
#> 
#> [[19]]$`1_month`
#> [1] 3270
#> 
#> [[19]]$`3_month`
#> NULL
#> 
#> [[19]]$`6_month`
#> NULL
#> 
#> [[19]]$`1_year`
#> NULL
#> 
#> [[19]]$other
#> [1] 3530
#> 
#> 
#> [[20]]
#> [[20]]$date
#> [1] "2016-02-21"
#> 
#> [[20]]$overnight
#> [1] 33369.02
#> 
#> [[20]]$`1_week`
#> [1] 500
#> 
#> [[20]]$`1_month`
#> [1] 3270
#> 
#> [[20]]$`3_month`
#> NULL
#> 
#> [[20]]$`6_month`
#> NULL
#> 
#> [[20]]$`1_year`
#> NULL
#> 
#> [[20]]$other
#> [1] 3530
#> 
#> 
#> [[21]]
#> [[21]]$date
#> [1] "2016-02-22"
#> 
#> [[21]]$overnight
#> [1] 35276.61
#> 
#> [[21]]$`1_week`
#> [1] 5140
#> 
#> [[21]]$`1_month`
#> NULL
#> 
#> [[21]]$`3_month`
#> NULL
#> 
#> [[21]]$`6_month`
#> NULL
#> 
#> [[21]]$`1_year`
#> NULL
#> 
#> [[21]]$other
#> [1] 2040
#> 
#> 
#> [[22]]
#> [[22]]$date
#> [1] "2016-02-23"
#> 
#> [[22]]$overnight
#> [1] 35150.78
#> 
#> [[22]]$`1_week`
#> [1] 625
#> 
#> [[22]]$`1_month`
#> NULL
#> 
#> [[22]]$`3_month`
#> NULL
#> 
#> [[22]]$`6_month`
#> NULL
#> 
#> [[22]]$`1_year`
#> NULL
#> 
#> [[22]]$other
#> [1] 209
#> 
#> 
#> [[23]]
#> [[23]]$date
#> [1] "2016-02-24"
#> 
#> [[23]]$overnight
#> [1] 37597.87
#> 
#> [[23]]$`1_week`
#> [1] 3720
#> 
#> [[23]]$`1_month`
#> NULL
#> 
#> [[23]]$`3_month`
#> NULL
#> 
#> [[23]]$`6_month`
#> NULL
#> 
#> [[23]]$`1_year`
#> NULL
#> 
#> [[23]]$other
#> NULL
#> 
#> 
#> [[24]]
#> [[24]]$date
#> [1] "2016-02-25"
#> 
#> [[24]]$overnight
#> [1] 40053.26
#> 
#> [[24]]$`1_week`
#> [1] 3700
#> 
#> [[24]]$`1_month`
#> [1] 150
#> 
#> [[24]]$`3_month`
#> NULL
#> 
#> [[24]]$`6_month`
#> NULL
#> 
#> [[24]]$`1_year`
#> NULL
#> 
#> [[24]]$other
#> [1] 990
#> 
#> 
#> [[25]]
#> [[25]]$date
#> [1] "2016-02-26"
#> 
#> [[25]]$overnight
#> [1] 37383.7
#> 
#> [[25]]$`1_week`
#> [1] 4045
#> 
#> [[25]]$`1_month`
#> [1] 100
#> 
#> [[25]]$`3_month`
#> NULL
#> 
#> [[25]]$`6_month`
#> NULL
#> 
#> [[25]]$`1_year`
#> NULL
#> 
#> [[25]]$other
#> [1] 430
#> 
#> 
#> [[26]]
#> [[26]]$date
#> [1] "2016-02-27"
#> 
#> [[26]]$overnight
#> [1] 37383.7
#> 
#> [[26]]$`1_week`
#> [1] 4045
#> 
#> [[26]]$`1_month`
#> [1] 100
#> 
#> [[26]]$`3_month`
#> NULL
#> 
#> [[26]]$`6_month`
#> NULL
#> 
#> [[26]]$`1_year`
#> NULL
#> 
#> [[26]]$other
#> [1] 430
#> 
#> 
#> [[27]]
#> [[27]]$date
#> [1] "2016-02-28"
#> 
#> [[27]]$overnight
#> [1] 37383.7
#> 
#> [[27]]$`1_week`
#> [1] 4045
#> 
#> [[27]]$`1_month`
#> [1] 100
#> 
#> [[27]]$`3_month`
#> NULL
#> 
#> [[27]]$`6_month`
#> NULL
#> 
#> [[27]]$`1_year`
#> NULL
#> 
#> [[27]]$other
#> [1] 430
#> 
#> 
#> [[28]]
#> [[28]]$date
#> [1] "2016-02-29"
#> 
#> [[28]]$overnight
#> [1] 38087.03
#> 
#> [[28]]$`1_week`
#> [1] 4490
#> 
#> [[28]]$`1_month`
#> [1] 150
#> 
#> [[28]]$`3_month`
#> NULL
#> 
#> [[28]]$`6_month`
#> NULL
#> 
#> [[28]]$`1_year`
#> NULL
#> 
#> [[28]]$other
#> [1] 300
#> 
#> 
if (FALSE) get_islamic_interbank_rate() # \dontrun{}
get_islamic_interbank_rate(date = "2018-01-01")
#> $date
#> [1] "2018-01-01"
#> 
#> $overnight
#> [1] 2.99
#> 
#> $`1_week`
#> NULL
#> 
#> $`1_month`
#> NULL
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 
get_islamic_interbank_rate(year = 2016, month = 2)
#> [[1]]
#> [[1]]$date
#> [1] "2016-02-01"
#> 
#> [[1]]$overnight
#> [1] 3.2
#> 
#> [[1]]$`1_week`
#> NULL
#> 
#> [[1]]$`1_month`
#> NULL
#> 
#> [[1]]$`3_month`
#> [1] 3.88
#> 
#> [[1]]$`6_month`
#> NULL
#> 
#> [[1]]$`1_year`
#> NULL
#> 
#> 
#> [[2]]
#> [[2]]$date
#> [1] "2016-02-03"
#> 
#> [[2]]$overnight
#> [1] 3.19
#> 
#> [[2]]$`1_week`
#> NULL
#> 
#> [[2]]$`1_month`
#> NULL
#> 
#> [[2]]$`3_month`
#> [1] 3.87
#> 
#> [[2]]$`6_month`
#> NULL
#> 
#> [[2]]$`1_year`
#> NULL
#> 
#> 
#> [[3]]
#> [[3]]$date
#> [1] "2016-02-04"
#> 
#> [[3]]$overnight
#> [1] 3.19
#> 
#> [[3]]$`1_week`
#> NULL
#> 
#> [[3]]$`1_month`
#> NULL
#> 
#> [[3]]$`3_month`
#> NULL
#> 
#> [[3]]$`6_month`
#> NULL
#> 
#> [[3]]$`1_year`
#> NULL
#> 
#> 
#> [[4]]
#> [[4]]$date
#> [1] "2016-02-05"
#> 
#> [[4]]$overnight
#> [1] 3.19
#> 
#> [[4]]$`1_week`
#> NULL
#> 
#> [[4]]$`1_month`
#> NULL
#> 
#> [[4]]$`3_month`
#> NULL
#> 
#> [[4]]$`6_month`
#> NULL
#> 
#> [[4]]$`1_year`
#> NULL
#> 
#> 
#> [[5]]
#> [[5]]$date
#> [1] "2016-02-06"
#> 
#> [[5]]$overnight
#> [1] 3.19
#> 
#> [[5]]$`1_week`
#> NULL
#> 
#> [[5]]$`1_month`
#> NULL
#> 
#> [[5]]$`3_month`
#> NULL
#> 
#> [[5]]$`6_month`
#> NULL
#> 
#> [[5]]$`1_year`
#> NULL
#> 
#> 
#> [[6]]
#> [[6]]$date
#> [1] "2016-02-07"
#> 
#> [[6]]$overnight
#> [1] 3.19
#> 
#> [[6]]$`1_week`
#> NULL
#> 
#> [[6]]$`1_month`
#> NULL
#> 
#> [[6]]$`3_month`
#> NULL
#> 
#> [[6]]$`6_month`
#> NULL
#> 
#> [[6]]$`1_year`
#> NULL
#> 
#> 
#> [[7]]
#> [[7]]$date
#> [1] "2016-02-08"
#> 
#> [[7]]$overnight
#> [1] 3.19
#> 
#> [[7]]$`1_week`
#> NULL
#> 
#> [[7]]$`1_month`
#> NULL
#> 
#> [[7]]$`3_month`
#> NULL
#> 
#> [[7]]$`6_month`
#> NULL
#> 
#> [[7]]$`1_year`
#> NULL
#> 
#> 
#> [[8]]
#> [[8]]$date
#> [1] "2016-02-10"
#> 
#> [[8]]$overnight
#> [1] 3.2
#> 
#> [[8]]$`1_week`
#> NULL
#> 
#> [[8]]$`1_month`
#> NULL
#> 
#> [[8]]$`3_month`
#> NULL
#> 
#> [[8]]$`6_month`
#> NULL
#> 
#> [[8]]$`1_year`
#> NULL
#> 
#> 
#> [[9]]
#> [[9]]$date
#> [1] "2016-02-11"
#> 
#> [[9]]$overnight
#> [1] 3.19
#> 
#> [[9]]$`1_week`
#> NULL
#> 
#> [[9]]$`1_month`
#> NULL
#> 
#> [[9]]$`3_month`
#> NULL
#> 
#> [[9]]$`6_month`
#> NULL
#> 
#> [[9]]$`1_year`
#> NULL
#> 
#> 
#> [[10]]
#> [[10]]$date
#> [1] "2016-02-12"
#> 
#> [[10]]$overnight
#> [1] 3.19
#> 
#> [[10]]$`1_week`
#> NULL
#> 
#> [[10]]$`1_month`
#> NULL
#> 
#> [[10]]$`3_month`
#> [1] 3.76
#> 
#> [[10]]$`6_month`
#> NULL
#> 
#> [[10]]$`1_year`
#> NULL
#> 
#> 
#> [[11]]
#> [[11]]$date
#> [1] "2016-02-13"
#> 
#> [[11]]$overnight
#> [1] 3.19
#> 
#> [[11]]$`1_week`
#> NULL
#> 
#> [[11]]$`1_month`
#> NULL
#> 
#> [[11]]$`3_month`
#> [1] 3.76
#> 
#> [[11]]$`6_month`
#> NULL
#> 
#> [[11]]$`1_year`
#> NULL
#> 
#> 
#> [[12]]
#> [[12]]$date
#> [1] "2016-02-14"
#> 
#> [[12]]$overnight
#> [1] 3.19
#> 
#> [[12]]$`1_week`
#> NULL
#> 
#> [[12]]$`1_month`
#> NULL
#> 
#> [[12]]$`3_month`
#> [1] 3.76
#> 
#> [[12]]$`6_month`
#> NULL
#> 
#> [[12]]$`1_year`
#> NULL
#> 
#> 
#> [[13]]
#> [[13]]$date
#> [1] "2016-02-15"
#> 
#> [[13]]$overnight
#> [1] 3.19
#> 
#> [[13]]$`1_week`
#> NULL
#> 
#> [[13]]$`1_month`
#> NULL
#> 
#> [[13]]$`3_month`
#> NULL
#> 
#> [[13]]$`6_month`
#> NULL
#> 
#> [[13]]$`1_year`
#> NULL
#> 
#> 
#> [[14]]
#> [[14]]$date
#> [1] "2016-02-17"
#> 
#> [[14]]$overnight
#> [1] 3.19
#> 
#> [[14]]$`1_week`
#> NULL
#> 
#> [[14]]$`1_month`
#> NULL
#> 
#> [[14]]$`3_month`
#> [1] 3.72
#> 
#> [[14]]$`6_month`
#> NULL
#> 
#> [[14]]$`1_year`
#> NULL
#> 
#> 
#> [[15]]
#> [[15]]$date
#> [1] "2016-02-18"
#> 
#> [[15]]$overnight
#> [1] 3.19
#> 
#> [[15]]$`1_week`
#> NULL
#> 
#> [[15]]$`1_month`
#> NULL
#> 
#> [[15]]$`3_month`
#> NULL
#> 
#> [[15]]$`6_month`
#> NULL
#> 
#> [[15]]$`1_year`
#> NULL
#> 
#> 
#> [[16]]
#> [[16]]$date
#> [1] "2016-02-19"
#> 
#> [[16]]$overnight
#> [1] 3.19
#> 
#> [[16]]$`1_week`
#> NULL
#> 
#> [[16]]$`1_month`
#> NULL
#> 
#> [[16]]$`3_month`
#> NULL
#> 
#> [[16]]$`6_month`
#> NULL
#> 
#> [[16]]$`1_year`
#> NULL
#> 
#> 
#> [[17]]
#> [[17]]$date
#> [1] "2016-02-20"
#> 
#> [[17]]$overnight
#> [1] 3.19
#> 
#> [[17]]$`1_week`
#> NULL
#> 
#> [[17]]$`1_month`
#> NULL
#> 
#> [[17]]$`3_month`
#> NULL
#> 
#> [[17]]$`6_month`
#> NULL
#> 
#> [[17]]$`1_year`
#> NULL
#> 
#> 
#> [[18]]
#> [[18]]$date
#> [1] "2016-02-21"
#> 
#> [[18]]$overnight
#> [1] 3.19
#> 
#> [[18]]$`1_week`
#> NULL
#> 
#> [[18]]$`1_month`
#> NULL
#> 
#> [[18]]$`3_month`
#> NULL
#> 
#> [[18]]$`6_month`
#> NULL
#> 
#> [[18]]$`1_year`
#> NULL
#> 
#> 
#> [[19]]
#> [[19]]$date
#> [1] "2016-02-22"
#> 
#> [[19]]$overnight
#> [1] 3.19
#> 
#> [[19]]$`1_week`
#> NULL
#> 
#> [[19]]$`1_month`
#> NULL
#> 
#> [[19]]$`3_month`
#> [1] 3.72
#> 
#> [[19]]$`6_month`
#> NULL
#> 
#> [[19]]$`1_year`
#> NULL
#> 
#> 
#> [[20]]
#> [[20]]$date
#> [1] "2016-02-23"
#> 
#> [[20]]$overnight
#> [1] 3.2
#> 
#> [[20]]$`1_week`
#> NULL
#> 
#> [[20]]$`1_month`
#> NULL
#> 
#> [[20]]$`3_month`
#> NULL
#> 
#> [[20]]$`6_month`
#> NULL
#> 
#> [[20]]$`1_year`
#> NULL
#> 
#> 
#> [[21]]
#> [[21]]$date
#> [1] "2016-02-24"
#> 
#> [[21]]$overnight
#> [1] 3.19
#> 
#> [[21]]$`1_week`
#> NULL
#> 
#> [[21]]$`1_month`
#> NULL
#> 
#> [[21]]$`3_month`
#> [1] 3.71
#> 
#> [[21]]$`6_month`
#> NULL
#> 
#> [[21]]$`1_year`
#> NULL
#> 
#> 
#> [[22]]
#> [[22]]$date
#> [1] "2016-02-25"
#> 
#> [[22]]$overnight
#> [1] 3.2
#> 
#> [[22]]$`1_week`
#> NULL
#> 
#> [[22]]$`1_month`
#> NULL
#> 
#> [[22]]$`3_month`
#> [1] 3.71
#> 
#> [[22]]$`6_month`
#> NULL
#> 
#> [[22]]$`1_year`
#> NULL
#> 
#> 
#> [[23]]
#> [[23]]$date
#> [1] "2016-02-26"
#> 
#> [[23]]$overnight
#> [1] 3.2
#> 
#> [[23]]$`1_week`
#> NULL
#> 
#> [[23]]$`1_month`
#> NULL
#> 
#> [[23]]$`3_month`
#> [1] 3.72
#> 
#> [[23]]$`6_month`
#> NULL
#> 
#> [[23]]$`1_year`
#> NULL
#> 
#> 
#> [[24]]
#> [[24]]$date
#> [1] "2016-02-27"
#> 
#> [[24]]$overnight
#> [1] 3.2
#> 
#> [[24]]$`1_week`
#> NULL
#> 
#> [[24]]$`1_month`
#> NULL
#> 
#> [[24]]$`3_month`
#> [1] 3.72
#> 
#> [[24]]$`6_month`
#> NULL
#> 
#> [[24]]$`1_year`
#> NULL
#> 
#> 
#> [[25]]
#> [[25]]$date
#> [1] "2016-02-28"
#> 
#> [[25]]$overnight
#> [1] 3.2
#> 
#> [[25]]$`1_week`
#> NULL
#> 
#> [[25]]$`1_month`
#> NULL
#> 
#> [[25]]$`3_month`
#> [1] 3.72
#> 
#> [[25]]$`6_month`
#> NULL
#> 
#> [[25]]$`1_year`
#> NULL
#> 
#> 
#> [[26]]
#> [[26]]$date
#> [1] "2016-02-29"
#> 
#> [[26]]$overnight
#> [1] 3.21
#> 
#> [[26]]$`1_week`
#> NULL
#> 
#> [[26]]$`1_month`
#> NULL
#> 
#> [[26]]$`3_month`
#> NULL
#> 
#> [[26]]$`6_month`
#> NULL
#> 
#> [[26]]$`1_year`
#> NULL
#> 
#>