Interest Volume

This function allows you to obtain daily interbank money market rates and volumes of transactions according to tenure. (2015 - present) from the BNM API.

interest_volume(product = "money_market_operations", date, year, month)

Arguments

product
Either "money_market_operations", "interbank" or "overall"

date
Character string of date with format as defined by RFC 3339, section 5.6 (YYYY-MM-DD). If specified, return values for the specified date.

year
Year and month as integers. If date, year and month left blank, return today's values.

month
Year and month as integers. If date, year and month left blank, return today's values.

Source

https://api.bnm.gov.my/

Examples

interest_volume()

#> $date
#> [1] "2019-05-01"
#> 
#> $overnight
#> [1] 3.174e+10
#> 
#> $`1_week`
#> [1] 1220000000
#> 
#> $`1_month`
#> NULL
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 
#> $other
#> [1] 1740000000
#> 

interest_volume(date = "2018-01-01")

#> $date
#> [1] "2018-01-01"
#> 
#> $overnight
#> [1] 3.31e+10
#> 
#> $`1_week`
#> [1] 1430000000
#> 
#> $`1_month`
#> [1] 200000000
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 
#> $other
#> [1] 5.03e+09
#> 

interest_volume(year = 2016, month = 2)

#>          date  overnight   1_week   1_month 3_month 6_month 1_year     other
#> 1  2016-02-01 3.2000e+10 3.86e+09        NA      NA      NA     NA 3.140e+09
#> 2  2016-02-02 3.2000e+10 1.86e+09 7.030e+09      NA      NA     NA 4.355e+09
#> 3  2016-02-03 3.2000e+10       NA        NA      NA      NA     NA 4.180e+09
#> 4  2016-02-04 3.2000e+10 3.00e+09        NA      NA      NA     NA        NA
#> 5  2016-02-05 3.0000e+10       NA        NA      NA      NA     NA 3.050e+09
#> 6  2016-02-06 3.0000e+10       NA        NA      NA      NA     NA 3.050e+09
#> 7  2016-02-07 3.0000e+10       NA        NA      NA      NA     NA 3.050e+09
#> 8  2016-02-08 3.0000e+10       NA        NA      NA      NA     NA 3.050e+09
#> 9  2016-02-09 3.0000e+10       NA        NA      NA      NA     NA 3.050e+09
#> 10 2016-02-10 3.0000e+10 3.80e+09        NA      NA      NA     NA 3.280e+09
#> 11 2016-02-11 3.2000e+10 1.50e+09        NA      NA      NA     NA 1.000e+09
#> 12 2016-02-12 3.1000e+10 1.70e+08 4.165e+09      NA      NA     NA 2.030e+09
#> 13 2016-02-13 3.1000e+10 1.70e+08 4.165e+09      NA      NA     NA 2.030e+09
#> 14 2016-02-14 3.1000e+10 1.70e+08 4.165e+09      NA      NA     NA 2.030e+09
#> 15 2016-02-15 3.0000e+10 2.00e+09        NA      NA      NA     NA        NA
#> 16 2016-02-16 3.2000e+10 2.00e+09        NA      NA      NA     NA        NA
#> 17 2016-02-17 3.1000e+10 2.00e+09        NA      NA      NA     NA        NA
#> 18 2016-02-18 3.3000e+10 4.00e+09        NA      NA      NA     NA        NA
#> 19 2016-02-19 3.1000e+10 4.50e+08 3.270e+09      NA      NA     NA 3.430e+09
#> 20 2016-02-20 3.1000e+10 4.50e+08 3.270e+09      NA      NA     NA 3.430e+09
#> 21 2016-02-21 3.1000e+10 4.50e+08 3.270e+09      NA      NA     NA 3.430e+09
#> 22 2016-02-22 3.2000e+10 4.94e+09        NA      NA      NA     NA 2.000e+09
#> 23 2016-02-23 3.3000e+10 2.88e+09        NA      NA      NA     NA 2.700e+09
#> 24 2016-02-24 3.6000e+10 3.62e+09        NA      NA      NA     NA        NA
#> 25 2016-02-25 3.8000e+10 3.50e+09        NA      NA      NA     NA        NA
#> 26 2016-02-26 3.2692e+10 4.00e+09        NA      NA      NA     NA        NA
#> 27 2016-02-27 3.2692e+10 4.00e+09        NA      NA      NA     NA        NA
#> 28 2016-02-28 3.2692e+10 4.00e+09        NA      NA      NA     NA        NA
#> 29 2016-02-29 3.4000e+10 4.33e+09        NA      NA      NA     NA        NA

interest_volume(product = "overall", year = 2016, month = 2)

#>          date   overnight    1_week   1_month 3_month 6_month 1_year     other
#> 1  2016-02-01 34068130000 3.860e+09 1.500e+08      NA      NA     NA 3.362e+09
#> 2  2016-02-02 33708700000 2.160e+09 7.110e+09      NA      NA     NA 4.355e+09
#> 3  2016-02-03 34533360000 1.550e+08 5.000e+07      NA      NA     NA 4.180e+09
#> 4  2016-02-04 34696380000 3.290e+09        NA      NA      NA     NA 1.500e+08
#> 5  2016-02-05 32216190000 2.000e+08        NA      NA      NA     NA 3.750e+09
#> 6  2016-02-06 32216190000 2.000e+08        NA      NA      NA     NA 3.750e+09
#> 7  2016-02-07 32216190000 2.000e+08        NA      NA      NA     NA 3.750e+09
#> 8  2016-02-08 32216190000 2.000e+08        NA      NA      NA     NA 3.750e+09
#> 9  2016-02-10 32485550000 3.900e+09        NA      NA      NA     NA 3.530e+09
#> 10 2016-02-11 34464270000 2.500e+09        NA      NA      NA     NA 1.250e+09
#> 11 2016-02-12 32944470000 1.700e+08 4.165e+09      NA      NA     NA 2.430e+09
#> 12 2016-02-13 32944470000 1.700e+08 4.165e+09      NA      NA     NA 2.430e+09
#> 13 2016-02-14 32944470000 1.700e+08 4.165e+09      NA      NA     NA 2.430e+09
#> 14 2016-02-15 32324530000 2.190e+09 6.000e+07      NA      NA     NA        NA
#> 15 2016-02-16 33780160000 2.300e+09 3.380e+06      NA      NA     NA 4.000e+08
#> 16 2016-02-17 33542950000 2.000e+09        NA      NA      NA     NA 5.100e+08
#> 17 2016-02-18 34370640000 4.180e+09 5.000e+07      NA      NA     NA 1.000e+08
#> 18 2016-02-19 33369020000 5.000e+08 3.270e+09      NA      NA     NA 3.530e+09
#> 19 2016-02-20 33369020000 5.000e+08 3.270e+09      NA      NA     NA 3.530e+09
#> 20 2016-02-21 33369020000 5.000e+08 3.270e+09      NA      NA     NA 3.530e+09
#> 21 2016-02-22 35276610000 5.140e+09        NA      NA      NA     NA 2.040e+09
#> 22 2016-02-23 35150780000 6.250e+08        NA      NA      NA     NA 2.090e+08
#> 23 2016-02-24 37597870000 3.720e+09        NA      NA      NA     NA        NA
#> 24 2016-02-25 40053260000 3.700e+09 1.500e+08      NA      NA     NA 9.900e+08
#> 25 2016-02-26 37383700000 4.045e+09 1.000e+08      NA      NA     NA 4.300e+08
#> 26 2016-02-27 37383700000 4.045e+09 1.000e+08      NA      NA     NA 4.300e+08
#> 27 2016-02-28 37383700000 4.045e+09 1.000e+08      NA      NA     NA 4.300e+08
#> 28 2016-02-29 38087030000 4.490e+09 1.500e+08      NA      NA     NA 3.000e+08