Islamic Interbank Rate

This function allows you to obtain daily weighted average of Islamic interbank deposit rates for various tenures. (Jan 2015-present) from the BNM API.

islamic_interbank_rate(date = NULL, year = NULL, month = NULL)

Arguments

date
Character string of date with format as defined by RFC 3339, section 5.6 (YYYY-MM-DD). If specified, return values for the specified date.

year, month
Year and month as integers. If date, year and month left blank, return today's values.

Source

https://api.bnm.gov.my/

Examples

islamic_interbank_rate()

#> $date
#> [1] "2019-05-02"
#> 
#> $overnight
#> [1] 3.21
#> 
#> $`1_week`
#> NULL
#> 
#> $`1_month`
#> NULL
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 

islamic_interbank_rate(date = "2018-01-01")

#> $date
#> [1] "2018-01-01"
#> 
#> $overnight
#> [1] 2.99
#> 
#> $`1_week`
#> NULL
#> 
#> $`1_month`
#> NULL
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 

islamic_interbank_rate(year = 2016, month = 2)

#>          date overnight 1_week 1_month 3_month 6_month 1_year
#> 1  2016-02-01      3.20     NA      NA    3.88      NA     NA
#> 2  2016-02-03      3.19     NA      NA    3.87      NA     NA
#> 3  2016-02-04      3.19     NA      NA      NA      NA     NA
#> 4  2016-02-05      3.19     NA      NA      NA      NA     NA
#> 5  2016-02-06      3.19     NA      NA      NA      NA     NA
#> 6  2016-02-07      3.19     NA      NA      NA      NA     NA
#> 7  2016-02-08      3.19     NA      NA      NA      NA     NA
#> 8  2016-02-10      3.20     NA      NA      NA      NA     NA
#> 9  2016-02-11      3.19     NA      NA      NA      NA     NA
#> 10 2016-02-12      3.19     NA      NA    3.76      NA     NA
#> 11 2016-02-13      3.19     NA      NA    3.76      NA     NA
#> 12 2016-02-14      3.19     NA      NA    3.76      NA     NA
#> 13 2016-02-15      3.19     NA      NA      NA      NA     NA
#> 14 2016-02-17      3.19     NA      NA    3.72      NA     NA
#> 15 2016-02-18      3.19     NA      NA      NA      NA     NA
#> 16 2016-02-19      3.19     NA      NA      NA      NA     NA
#> 17 2016-02-20      3.19     NA      NA      NA      NA     NA
#> 18 2016-02-21      3.19     NA      NA      NA      NA     NA
#> 19 2016-02-22      3.19     NA      NA    3.72      NA     NA
#> 20 2016-02-23      3.20     NA      NA      NA      NA     NA
#> 21 2016-02-24      3.19     NA      NA    3.71      NA     NA
#> 22 2016-02-25      3.20     NA      NA    3.71      NA     NA
#> 23 2016-02-26      3.20     NA      NA    3.72      NA     NA
#> 24 2016-02-27      3.20     NA      NA    3.72      NA     NA
#> 25 2016-02-28      3.20     NA      NA    3.72      NA     NA
#> 26 2016-02-29      3.21     NA      NA      NA      NA     NA