USD Interbank Intraday Rate

This function allows you to obtain USD/MYR interbank intraday highest and lowest rate from the BNM API. Rates are obtained from the best U.S. dollar against Malaysian ringgit interbank highest and lowest dealt rates by commercial banks on the specific date.

usd_interbank_intraday_rate(date = NULL, year = NULL, month = NULL)

Arguments

date
Character string of date with format as defined by RFC 3339, section 5.6 (YYYY-MM-DD). If specified, return values for the specified date.

year
Year and month as integers. If date, year and month left blank, return today's values.

month
Year and month as integers. If date, year and month left blank, return today's values.

Source

https://api.bnm.gov.my/

Examples

usd_interbank_intraday_rate()

#> $date
#> [1] "2019-05-03"
#> 
#> $highest_rate
#> [1] 4.143
#> 
#> $lowest_rate
#> [1] 4.141
#> 

usd_interbank_intraday_rate(date = "2018-01-03")

#> $date
#> [1] "2018-01-03"
#> 
#> $highest_rate
#> [1] 4.028
#> 
#> $lowest_rate
#> [1] 4.014
#> 

usd_interbank_intraday_rate(year = 2016, month = 2)

#>          date highest_rate lowest_rate
#> 1  2016-02-02        4.215      4.1600
#> 2  2016-02-03        4.245      4.2220
#> 3  2016-02-04        4.165      4.1430
#> 4  2016-02-05        4.168      4.1170
#> 5  2016-02-10        4.180      4.1200
#> 6  2016-02-11        4.150      4.1005
#> 7  2016-02-12        4.190      4.1400
#> 8  2016-02-15        4.167      4.1155
#> 9  2016-02-16        4.156      4.1150
#> 10 2016-02-17        4.223      4.1900
#> 11 2016-02-18        4.181      4.1500
#> 12 2016-02-19        4.233      4.1850
#> 13 2016-02-22        4.218      4.1950
#> 14 2016-02-23        4.213      4.1730
#> 15 2016-02-24        4.247      4.2150
#> 16 2016-02-25        4.224      4.2025
#> 17 2016-02-26        4.232      4.1900
#> 18 2016-02-29        4.235      4.2030