Base rate

This function allows you to obtain current Base Rates and Base Lending Rates for retail loans or financing facilities and Indicative Effective Lending Rates for a standard housing loan/home financing offered by financial institutions from the BNM API.

base_rate(bank_code = NULL)

Arguments

bank_code
8 characters of SWIFT code per https://api.bnm.gov.my/

Source

https://api.bnm.gov.my/

Examples

base_rate()

#>    bank_code                                                 bank_name
#> 1   BKKBMYKL                                       Bangkok Bank Berhad
#> 2   CIBBMYKL                                          CIMB Bank Berhad
#> 3   CITIMYKL                                           Citibank Berhad
#> 4   HLBBMYKL                           Hong Leong Bank Malaysia Berhad
#> 5   HBMBMYKL                                 HSBC Bank Malaysia Berhad
#> 6   ICBKMYKL Industrial and Commercial Bank of China (Malaysia) Berhad
#> 7   MBBEMYKL                                    Malayan Banking Berhad
#> 8   OCBCMYKL                               OCBC Bank (Malaysia) Berhad
#> 9   PBBEMYKL                                        Public Bank Berhad
#> 10  RHBBMYKL                                           RHB Bank Berhad
#> 11  SCBLMYKX                   Standard Chartered Bank Malaysia Berhad
#> 12  UOVBMYKL                      United Overseas Bank (Malaysia) Bhd.
#> 13  AIBBMYKL                                 Affin Islamic Bank Berhad
#> 14  ALSRMYKL                              Alliance Islamic Bank Berhad
#> 15  AISLMYKL                                     AmBank Islamic Berhad
#> 16  BIMBMYKL                                Bank Islam Malaysia Berhad
#> 17  BMMBMYKL                             Bank Muamalat Malaysia Berhad
#> 18  CTBBMYKL                                  CIMB Islamic Bank Berhad
#> 19  HLIBMYKL                            Hong Leong Islamic Bank Berhad
#> 20  HMABMYKL                               HSBC Amanah Malaysia Berhad
#> 21  KFHOMYKL                    Kuwait Finance House (Malaysia) Berhad
#> 22  MBISMYKL                                    Maybank Islamic Berhad
#> 23  AFBQMYKL                                          MBSB Bank Berhad
#> 24  OABBMYKL                                  OCBC Al-Amin Bank Berhad
#> 25  PUIBMYKL                                Public Islamic Bank Berhad
#> 26  RHBAMYKL                                   RHB Islamic Bank Berhad
#> 27  SCSRMYKK                          Standard Chartered Saadiq Berhad
#> 28  AGOBMYKL                                                  Agrobank
#> 29  BSNAMYK1                                    Bank Simpanan Nasional
#> 30  PHBMMYKL                                         Affin Bank Berhad
#> 31  MFBBMYKL                             Alliance Bank Malaysia Berhad
#> 32  ARBKMYKL                                         AmBank (M) Berhad
#> 33  BKCHMYKL                                  Bank of China (M) Berhad
#> 34  RJHIMYKL                  Al Rajhi Banking & Investment (M) Berhad
#> 35  BKRMMYKL                          Bank Kerjasama Rakyat (M) Berhad
#>    base_rate base_lending_rate indicative_eff_lending_rate base_financing_rate
#> 1       4.72              7.37                        5.92                  NA
#> 2       4.25              7.10                        5.00                  NA
#> 3       3.90              7.05                        4.70                  NA
#> 4       4.13              7.14                        5.00                  NA
#> 5       3.89              6.99                        5.00                  NA
#> 6       4.02              6.95                        4.75                  NA
#> 7       3.25              6.90                        4.60                  NA
#> 8       4.08              7.01                        5.16                  NA
#> 9       3.77              6.97                        4.60                  NA
#> 10      4.00              6.95                        5.00                  NA
#> 11      3.77              6.95                        4.77                  NA
#> 12      4.11              7.07                        4.86                  NA
#> 13      4.20                NA                        5.01                7.06
#> 14      4.07                NA                        4.61                6.92
#> 15      4.10                NA                        4.75                6.95
#> 16      4.03                NA                        4.83                6.98
#> 17      4.06                NA                        5.06                7.06
#> 18      4.25                NA                        5.00                7.10
#> 19      4.13                NA                        4.85                7.14
#> 20      3.89                NA                        5.00                6.99
#> 21      3.75                NA                        4.60                7.64
#> 22      3.25                NA                        4.60                6.90
#> 23      4.15                NA                        4.80                7.00
#> 24      4.08                NA                        5.16                7.01
#> 25      3.77                NA                        4.60                6.97
#> 26      4.00                NA                        5.00                6.95
#> 27      3.77                NA                        4.77                6.95
#> 28      3.85                NA                          NA                7.00
#> 29      4.10                NA                        4.70                6.85
#> 30      4.20              7.06                        5.01                  NA
#> 31      4.07              6.92                        4.61                  NA
#> 32      4.10              6.95                        4.75                  NA
#> 33      4.05              6.85                        5.05                  NA
#> 34      4.35                NA                        5.70                7.25
#> 35      4.10                NA                        4.90                7.08