Base Rate

This function allows you to obtain current Base Rates and Base Lending Rates for retail loans or financing facilities and Indicative Effective Lending Rates for a standard housing loan/home financing offered by financial institutions from the BNM API.

get_base_rate(bank_code = NULL)

Arguments

bank_code
8 characters of SWIFT code per https://api.bnm.gov.my/

Source

https://api.bnm.gov.my/

Examples

get_base_rate()

#> # A tibble: 35 x 6
#>    bank_code   bank_name           base_rate base_lending_ra~ indicative_eff_le~
#>    <chr>       <chr>                   <dbl>            <dbl>              <dbl>
#>  1 "AFFINBANK" Affin Bank               2.7              5.56               3.3 
#>  2 "ALLIANCEB" Alliance Bank            2.57             5.42               3.11
#>  3 "AMBANK"    AmBank                   2.6              5.45               3.25
#>  4 "BANGKOK"   Bangkok Bank Berhad      3.22             5.87               4.42
#>  5 "BANKCHINA" Bank of China (Mal~      2.55             5.35               3.55
#>  6 "CIMB BANK" CIMB Bank Berhad         2.75             5.6                3.5 
#>  7 "CITIBANK"  Citibank Berhad          2.4              5.55               3.2 
#>  8 ""          Hong Leong Bank Ma~      2.63             5.64               3.5 
#>  9 ""          HSBC Bank Malaysia~      2.39             5.49               3.5 
#> 10 ""          Industrial and Com~      2.52             5.45               3.47
#> # ... with 25 more rows, and 1 more variable: base_financing_rate <dbl>