Malaysia Overnight Rate - I

Malaysia Overnight Rate - I

Usage

get_overnight_rate(date = NULL, year = NULL, month = NULL)

Source

https://apikijangportal.bnm.gov.my/

Arguments

date

Character string of date with format as defined by RFC 3339, section 5.6 (YYYY-MM-DD). If specified, return values for the specified date.

year, month

Year and month as integers. If date, year and month left blank, return today's values.

Examples

if (FALSE) get_overnight_rate()
get_overnight_rate(date = "2022-05-05")
#> [[1]]
#> [[1]]$reference_date
#> [1] "2022-05-05"
#> 
#> [[1]]$reference_rate
#> [1] "1.74"
#> 
#> [[1]]$aggregate_volume
#> [1] "23642.90"
#> 
#> 
get_overnight_rate(year = 2022, month = 5)
#> [[1]]
#> [[1]]$reference_date
#> [1] "2022-05-05"
#> 
#> [[1]]$reference_rate
#> [1] "1.74"
#> 
#> [[1]]$aggregate_volume
#> [1] "23642.90"
#> 
#> 
#> [[2]]
#> [[2]]$reference_date
#> [1] "2022-05-06"
#> 
#> [[2]]$reference_rate
#> [1] "1.75"
#> 
#> [[2]]$aggregate_volume
#> [1] "23977.60"
#> 
#> 
#> [[3]]
#> [[3]]$reference_date
#> [1] "2022-05-09"
#> 
#> [[3]]$reference_rate
#> [1] "1.73"
#> 
#> [[3]]$aggregate_volume
#> [1] "21970.30"
#> 
#> 
#> [[4]]
#> [[4]]$reference_date
#> [1] "2022-05-10"
#> 
#> [[4]]$reference_rate
#> [1] "1.69"
#> 
#> [[4]]$aggregate_volume
#> [1] "25134.10"
#> 
#> 
#> [[5]]
#> [[5]]$reference_date
#> [1] "2022-05-11"
#> 
#> [[5]]$reference_rate
#> [1] "1.94"
#> 
#> [[5]]$aggregate_volume
#> [1] "26474.70"
#> 
#> 
#> [[6]]
#> [[6]]$reference_date
#> [1] "2022-05-12"
#> 
#> [[6]]$reference_rate
#> [1] "1.99"
#> 
#> [[6]]$aggregate_volume
#> [1] "21897.30"
#> 
#> 
#> [[7]]
#> [[7]]$reference_date
#> [1] "2022-05-13"
#> 
#> [[7]]$reference_rate
#> [1] "1.96"
#> 
#> [[7]]$aggregate_volume
#> [1] "24911.10"
#> 
#> 
#> [[8]]
#> [[8]]$reference_date
#> [1] "2022-05-17"
#> 
#> [[8]]$reference_rate
#> [1] "1.92"
#> 
#> [[8]]$aggregate_volume
#> [1] "19829.00"
#> 
#> 
#> [[9]]
#> [[9]]$reference_date
#> [1] "2022-05-18"
#> 
#> [[9]]$reference_rate
#> [1] "1.95"
#> 
#> [[9]]$aggregate_volume
#> [1] "20096.50"
#> 
#> 
#> [[10]]
#> [[10]]$reference_date
#> [1] "2022-05-19"
#> 
#> [[10]]$reference_rate
#> [1] "1.96"
#> 
#> [[10]]$aggregate_volume
#> [1] "21513.40"
#> 
#> 
#> [[11]]
#> [[11]]$reference_date
#> [1] "2022-05-20"
#> 
#> [[11]]$reference_rate
#> [1] "1.96"
#> 
#> [[11]]$aggregate_volume
#> [1] "23137.70"
#> 
#> 
#> [[12]]
#> [[12]]$reference_date
#> [1] "2022-05-23"
#> 
#> [[12]]$reference_rate
#> [1] "1.96"
#> 
#> [[12]]$aggregate_volume
#> [1] "26742.90"
#> 
#> 
#> [[13]]
#> [[13]]$reference_date
#> [1] "2022-05-24"
#> 
#> [[13]]$reference_rate
#> [1] "1.99"
#> 
#> [[13]]$aggregate_volume
#> [1] "29034.20"
#> 
#> 
#> [[14]]
#> [[14]]$reference_date
#> [1] "2022-05-25"
#> 
#> [[14]]$reference_rate
#> [1] "1.99"
#> 
#> [[14]]$aggregate_volume
#> [1] "26342.60"
#> 
#> 
#> [[15]]
#> [[15]]$reference_date
#> [1] "2022-05-26"
#> 
#> [[15]]$reference_rate
#> [1] "2.00"
#> 
#> [[15]]$aggregate_volume
#> [1] "25301.60"
#> 
#> 
#> [[16]]
#> [[16]]$reference_date
#> [1] "2022-05-27"
#> 
#> [[16]]$reference_rate
#> [1] "2.00"
#> 
#> [[16]]$aggregate_volume
#> [1] "21929.30"
#> 
#> 
#> [[17]]
#> [[17]]$reference_date
#> [1] "2022-05-30"
#> 
#> [[17]]$reference_rate
#> [1] "2.00"
#> 
#> [[17]]$aggregate_volume
#> [1] "23067.94"
#> 
#> 
#> [[18]]
#> [[18]]$reference_date
#> [1] "2022-05-31"
#> 
#> [[18]]$reference_rate
#> [1] "2.00"
#> 
#> [[18]]$aggregate_volume
#> [1] "20133.90"
#> 
#>