Interest Rate

This function allows you to obtain daily interbank money market rates and volumes of transactions according to tenure. (2015 - present) from the BNM API.

interest_rate(product = "money_market_operations", date, year, month)

Arguments

product
Either "money_market_operations", "interbank" or "overall"

date
Character string of date with format as defined by RFC 3339, section 5.6 (YYYY-MM-DD). If specified, return values for the specified date.

year
Year and month as integers. If date, year and month left blank, return today's values.

month
Year and month as integers. If date, year and month left blank, return today's values.

Source

https://api.bnm.gov.my/

Examples

interest_rate()

#> $date
#> [1] "2019-05-01"
#> 
#> $overnight
#> [1] 3.24
#> 
#> $`1_week`
#> [1] 3.27
#> 
#> $`1_month`
#> NULL
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 

interest_rate(date = "2018-01-01")

#> $date
#> [1] "2018-01-01"
#> 
#> $overnight
#> [1] 2.91
#> 
#> $`1_week`
#> [1] 3.02
#> 
#> $`1_month`
#> [1] 3.08
#> 
#> $`3_month`
#> NULL
#> 
#> $`6_month`
#> NULL
#> 
#> $`1_year`
#> NULL
#> 

interest_rate(year = 2016, month = 2)

#>          date overnight 1_week 1_month 3_month 6_month 1_year
#> 1  2016-02-01      3.07   3.14      NA      NA      NA     NA
#> 2  2016-02-02      3.10   3.15    3.20      NA      NA     NA
#> 3  2016-02-03      3.23     NA      NA      NA      NA     NA
#> 4  2016-02-04      3.13   3.14      NA      NA      NA     NA
#> 5  2016-02-05      3.06     NA      NA      NA      NA     NA
#> 6  2016-02-06      3.06     NA      NA      NA      NA     NA
#> 7  2016-02-07      3.06     NA      NA      NA      NA     NA
#> 8  2016-02-08      3.06     NA      NA      NA      NA     NA
#> 9  2016-02-09      3.06     NA      NA      NA      NA     NA
#> 10 2016-02-10      3.09   3.14      NA      NA      NA     NA
#> 11 2016-02-11      3.05   3.12      NA      NA      NA     NA
#> 12 2016-02-12      3.11   3.13    3.20      NA      NA     NA
#> 13 2016-02-13      3.11   3.13    3.20      NA      NA     NA
#> 14 2016-02-14      3.11   3.13    3.20      NA      NA     NA
#> 15 2016-02-15      3.04   3.12      NA      NA      NA     NA
#> 16 2016-02-16      3.14   3.11      NA      NA      NA     NA
#> 17 2016-02-17      3.15   3.11      NA      NA      NA     NA
#> 18 2016-02-18      3.20   3.11      NA      NA      NA     NA
#> 19 2016-02-19      3.16   3.10    3.19      NA      NA     NA
#> 20 2016-02-20      3.16   3.10    3.19      NA      NA     NA
#> 21 2016-02-21      3.16   3.10    3.19      NA      NA     NA
#> 22 2016-02-22      3.16   3.12      NA      NA      NA     NA
#> 23 2016-02-23      3.18   3.11      NA      NA      NA     NA
#> 24 2016-02-24      3.19   3.13      NA      NA      NA     NA
#> 25 2016-02-25      3.10   3.13      NA      NA      NA     NA
#> 26 2016-02-26      3.24   3.14      NA      NA      NA     NA
#> 27 2016-02-27      3.24   3.14      NA      NA      NA     NA
#> 28 2016-02-28      3.24   3.14      NA      NA      NA     NA
#> 29 2016-02-29      3.19   3.15      NA      NA      NA     NA

interest_rate(product = "overall", year = 2016, month = 2)

#>          date overnight 1_week 1_month 3_month 6_month 1_year
#> 1  2016-02-01      3.07   3.14    3.50      NA      NA     NA
#> 2  2016-02-02      3.10   3.16    3.20      NA      NA     NA
#> 3  2016-02-03      3.22   3.24    3.36      NA      NA     NA
#> 4  2016-02-04      3.13   3.14      NA      NA      NA     NA
#> 5  2016-02-05      3.06   3.22      NA      NA      NA     NA
#> 6  2016-02-06      3.06   3.22      NA      NA      NA     NA
#> 7  2016-02-07      3.06   3.22      NA      NA      NA     NA
#> 8  2016-02-08      3.06   3.22      NA      NA      NA     NA
#> 9  2016-02-10      3.09   3.14      NA      NA      NA     NA
#> 10 2016-02-11      3.05   3.16      NA      NA      NA     NA
#> 11 2016-02-12      3.11   3.13    3.20      NA      NA     NA
#> 12 2016-02-13      3.11   3.13    3.20      NA      NA     NA
#> 13 2016-02-14      3.11   3.13    3.20      NA      NA     NA
#> 14 2016-02-15      3.04   3.12    3.36      NA      NA     NA
#> 15 2016-02-16      3.14   3.13    3.10      NA      NA     NA
#> 16 2016-02-17      3.14   3.11      NA      NA      NA     NA
#> 17 2016-02-18      3.20   3.11    3.37      NA      NA     NA
#> 18 2016-02-19      3.16   3.11    3.19      NA      NA     NA
#> 19 2016-02-20      3.16   3.11    3.19      NA      NA     NA
#> 20 2016-02-21      3.16   3.11    3.19      NA      NA     NA
#> 21 2016-02-22      3.15   3.12      NA      NA      NA     NA
#> 22 2016-02-23      3.18   3.23      NA      NA      NA     NA
#> 23 2016-02-24      3.19   3.13      NA      NA      NA     NA
#> 24 2016-02-25      3.10   3.13    3.37      NA      NA     NA
#> 25 2016-02-26      3.23   3.14    3.38      NA      NA     NA
#> 26 2016-02-27      3.23   3.14    3.38      NA      NA     NA
#> 27 2016-02-28      3.23   3.14    3.38      NA      NA     NA
#> 28 2016-02-29      3.19   3.16    3.41      NA      NA     NA