Kijang Emas

This function allows you to obtain daily trading prices of Malaysia gold bullion coin from the BNM API.

kijang_emas(date = NULL, year = NULL, month = NULL)

Arguments

date
Character string of date with format as defined by RFC 3339, section 5.6 (YYYY-MM-DD). If specified, return values for the specified date.

year
Year and month as integers. If date, year and month left blank, return today's values.

month
Year and month as integers. If date, year and month left blank, return today's values.

Source

https://api.bnm.gov.my/

Examples

kijang_emas()

#> $effective_date
#> [1] "2019-05-03"
#> 
#> $one_oz
#> $one_oz$buying
#> [1] 5384
#> 
#> $one_oz$selling
#> [1] 5599
#> 
#> 
#> $half_oz
#> $half_oz$buying
#> [1] 2692
#> 
#> $half_oz$selling
#> [1] 2852
#> 
#> 
#> $quarter_oz
#> $quarter_oz$buying
#> [1] 1346
#> 
#> $quarter_oz$selling
#> [1] 1453
#> 
#> 

kijang_emas(date = "2018-01-03")

#> $effective_date
#> [1] "2018-01-03"
#> 
#> $one_oz
#> $one_oz$buying
#> [1] 5415
#> 
#> $one_oz$selling
#> [1] 5632
#> 
#> 
#> $half_oz
#> $half_oz$buying
#> [1] 2708
#> 
#> $half_oz$selling
#> [1] 2869
#> 
#> 
#> $quarter_oz
#> $quarter_oz$buying
#> [1] 1354
#> 
#> $quarter_oz$selling
#> [1] 1461
#> 
#> 

kijang_emas(year = 2016, month = 2)

#>    effective_date one_oz.buying one_oz.selling half_oz.buying half_oz.selling
#> 1      2016-02-02          4790           4999           2395            2547
#> 2      2016-02-04          4836           5044           2418            2570
#> 3      2016-02-05          4843           5054           2422            2575
#> 4      2016-02-10          5052           5271           2526            2685
#> 5      2016-02-11          5060           5283           2530            2691
#> 6      2016-02-12          5243           5473           2622            2788
#> 7      2016-02-15          5215           5441           2607            2772
#> 8      2016-02-16          5079           5302           2540            2701
#> 9      2016-02-17          5130           5353           2565            2727
#> 10     2016-02-18          5132           5356           2566            2728
#> 11     2016-02-19          5235           5463           2618            2783
#> 12     2016-02-22          5257           5484           2629            2794
#> 13     2016-02-23          5148           5371           2574            2736
#> 14     2016-02-24          5279           5514           2640            2809
#> 15     2016-02-25          5265           5496           2633            2800
#> 16     2016-02-26          5270           5509           2635            2807
#> 17     2016-02-29          5255           5485           2627            2794
#>    quarter_oz.buying quarter_oz.selling
#> 1               1198               1297
#> 2               1209               1309
#> 3               1211               1311
#> 4               1263               1367
#> 5               1265               1371
#> 6               1311               1420
#> 7               1304               1412
#> 8               1270               1376
#> 9               1282               1389
#> 10              1283               1389
#> 11              1309               1417
#> 12              1314               1423
#> 13              1287               1394
#> 14              1320               1431
#> 15              1316               1426
#> 16              1318               1429
#> 17              1314               1423